@@ -1,41 +1,52 @@
 module github.com/txthinking/brook
 
-go 1.19
+go 1.21.5
 
 require (
-	github.com/gorilla/mux v1.8.0
-	github.com/gorilla/websocket v1.4.2
-	github.com/miekg/dns v1.1.43
+	github.com/gorilla/mux v1.8.1
+	github.com/gorilla/websocket v1.5.1
+	github.com/krolaw/dhcp4 v0.0.0-20190909130307-a50d88189771
+	github.com/miekg/dns v1.1.57
 	github.com/patrickmn/go-cache v2.1.0+incompatible
-	github.com/phuslu/iploc v1.0.20230201
-	github.com/quic-go/quic-go v0.32.0
+	github.com/phuslu/iploc v1.0.20240501
+	github.com/prometheus/client_golang v1.17.0
+	github.com/quic-go/quic-go v0.42.0
+	github.com/refraction-networking/utls v1.5.4
 	github.com/tdewolff/minify v2.3.6+incompatible
-	github.com/txthinking/crypto v0.0.0-20210716135230-de9624a415a4
-	github.com/txthinking/runnergroup v0.0.0-20230211072751-d11f16258c86
-	github.com/txthinking/socks5 v0.0.0-20230215122820-1d06f3a253de
-	github.com/txthinking/x v0.0.0-20220929041811-1b4d914e9133
-	github.com/urfave/cli/v2 v2.23.0
+	github.com/txthinking/runnergroup v0.0.0-20230325130830-408dc5853f86
+	github.com/txthinking/socks5 v0.0.0-20230325130024-4230056ae301
+	github.com/txthinking/x v0.0.0-20240301021728-6f68aba84c87
+	github.com/urfave/cli/v2 v2.25.7
 	github.com/urfave/negroni v1.0.0
-	golang.org/x/crypto v0.4.0
-	golang.org/x/net v0.7.0
-	golang.org/x/sys v0.5.0
+	golang.org/x/crypto v0.22.0
+	golang.org/x/net v0.24.0
 )
 
 require (
+	github.com/andybalholm/brotli v1.0.5 // indirect
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/cloudflare/circl v1.3.7 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
-	github.com/go-task/slim-sprig v0.0.0-20210107165309-348f09dbbbc0 // indirect
-	github.com/golang/mock v1.6.0 // indirect
+	github.com/gaukas/godicttls v0.0.4 // indirect
+	github.com/go-task/slim-sprig v0.0.0-20230315185526-52ccab3ef572 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
 	github.com/google/pprof v0.0.0-20210407192527-94a9f03dee38 // indirect
-	github.com/onsi/ginkgo/v2 v2.2.0 // indirect
-	github.com/quic-go/qtls-go1-18 v0.2.0 // indirect
-	github.com/quic-go/qtls-go1-19 v0.2.0 // indirect
-	github.com/quic-go/qtls-go1-20 v0.1.0 // indirect
+	github.com/klauspost/compress v1.16.7 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/onsi/ginkgo/v2 v2.9.5 // indirect
+	github.com/prometheus/client_model v0.4.1-0.20230718164431-9a2bf3000d16 // indirect
+	github.com/prometheus/common v0.44.0 // indirect
+	github.com/prometheus/procfs v0.11.1 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/tdewolff/parse v2.3.4+incompatible // indirect
-	github.com/tdewolff/test v1.0.6 // indirect
+	github.com/tdewolff/test v1.0.10 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	go.uber.org/mock v0.4.0 // indirect
 	golang.org/x/exp v0.0.0-20221205204356-47842c84f3db // indirect
-	golang.org/x/mod v0.6.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
-	golang.org/x/tools v0.2.0 // indirect
+	golang.org/x/mod v0.12.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/tools v0.13.0 // indirect
+	google.golang.org/protobuf v1.33.0 // indirect
 )
