@@ -15,12 +15,8 @@
 package main
 
 import (
-	"context"
-	"crypto/x509"
 	"errors"
 	"fmt"
-	"io/fs"
-	"io/ioutil"
 	"log"
 	"net"
 	"os"
@@ -28,20 +24,25 @@ import (
 	"os/signal"
 	"path/filepath"
 	"runtime"
+	"strconv"
 	"strings"
-	"sync"
 	"syscall"
 	"time"
 
-	"net/http"
 	"net/url"
 
+	"github.com/miekg/dns"
+	"github.com/phuslu/iploc"
 	"github.com/txthinking/brook"
 	"github.com/txthinking/brook/plugins/block"
+	"github.com/txthinking/brook/plugins/dialwithdns"
+	"github.com/txthinking/brook/plugins/dialwithip"
+	"github.com/txthinking/brook/plugins/dialwithnic"
+	"github.com/txthinking/brook/plugins/logger"
 	"github.com/txthinking/brook/plugins/pprof"
+	"github.com/txthinking/brook/plugins/prometheus"
 	"github.com/txthinking/brook/plugins/socks5dial"
 	"github.com/txthinking/brook/plugins/thedns"
-	"github.com/txthinking/brook/plugins/tproxy"
 	"github.com/txthinking/runnergroup"
 	"github.com/txthinking/socks5"
 	"github.com/urfave/cli/v2"
@@ -49,10 +50,11 @@ import (
 
 func main() {
 	g := runnergroup.New()
+	df := func() {}
 	app := cli.NewApp()
 	app.Name = "Brook"
-	app.Version = "20230218"
-	app.Usage = "A cross-platform network tool designed for developers"
+	app.Version = "20240606"
+	app.Usage = "A cross-platform programmable network tool"
 	app.Authors = []*cli.Author{
 		{
 			Name:  "Cloud",
@@ -66,8 +68,110 @@ func main() {
 			Name:  "pprof",
 			Usage: "go http pprof listen addr, such as :6060",
 		},
+		&cli.StringFlag{
+			Name:  "log",
+			Usage: "Enable log. A valid value is file path or 'console'. Send SIGUSR1 to me to reset the log file on unix system. If you want to debug SOCKS5 lib, set env SOCKS5_DEBUG=true",
+		},
+		&cli.StringSliceFlag{
+			Name:  "tag",
+			Usage: "Tag can be used to the process, will be append into log or serverLog, such as: 'key1:value1'. All tags will also be appended as query parameters one by one to the userAPI",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithDNS",
+			Usage: "When a domain name needs to be resolved, use the specified DNS. Such as 8.8.8.8:53 or https://dns.google/dns-query?address=8.8.8.8%3A443, the address is required. Note that for client-side commands, this does not affect the client passing the domain address to the server",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithDNSPrefer",
+			Usage: "This is used with the dialWithDNS parameter. Prefer A record or AAAA record. Value is A or AAAA",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithIP4",
+			Usage: "When the current machine establishes a network connection to the outside IPv4, both TCP and UDP, it is used to specify the IPv4 used",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithIP6",
+			Usage: "When the current machine establishes a network connection to the outside IPv6, both TCP and UDP, it is used to specify the IPv6 used",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithNIC",
+			Usage: "When the current machine establishes a network connection to the outside, both TCP and UDP, it is used to specify the NIC used",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithSocks5",
+			Usage: "When the current machine establishes a network connection to the outside, both TCP and UDP, with your socks5 proxy, such as 127.0.0.1:1081",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithSocks5Username",
+			Usage: "If there is",
+		},
+		&cli.StringFlag{
+			Name:  "dialWithSocks5Password",
+			Usage: "If there is",
+		},
+		&cli.IntFlag{
+			Name:  "dialWithSocks5TCPTimeout",
+			Value: 0,
+			Usage: "time (s)",
+		},
+		&cli.IntFlag{
+			Name:  "dialWithSocks5UDPTimeout",
+			Value: 60,
+			Usage: "time (s)",
+		},
+		&cli.StringFlag{
+			Name:  "prometheus",
+			Usage: "prometheus http listen addr, such as :7070. If it is transmitted on the public network, it is recommended to use it with nico",
+		},
+		&cli.StringFlag{
+			Name:  "prometheusPath",
+			Usage: "prometheus http path, such as /xxx. If it is transmitted on the public network, a hard-to-guess value is recommended",
+		},
+		&cli.StringFlag{
+			Name:  "clientHKDFInfo",
+			Usage: "client HKDF info, most time you don't need to change this, if changed, all and each brook links in client side must be same, I mean each",
+			Value: "brook",
+		},
+		&cli.StringFlag{
+			Name:  "serverHKDFInfo",
+			Usage: "server HKDF info, most time you don't need to change this, if changed, all and each brook links in client side must be same, I mean each",
+			Value: "brook",
+		},
+		&cli.StringFlag{
+			Name:  "blockDomainList",
+			Usage: "One domain per line, suffix match mode. https://, http:// or local file absolute path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt. Works with server/wsserver/wssserver/quicserver",
+		},
+		&cli.StringFlag{
+			Name:  "blockCIDR4List",
+			Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_cidr4.txt. Works with server/wsserver/wssserver/quicserver",
+		},
+		&cli.StringFlag{
+			Name:  "blockCIDR6List",
+			Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_cidr6.txt. Works with server/wsserver/wssserver/quicserver",
+		},
+		&cli.StringSliceFlag{
+			Name:  "blockGeoIP",
+			Usage: "Block IP by Geo country code, such as US. Works with server/wsserver/wssserver/quicserver",
+		},
+		&cli.Int64Flag{
+			Name:  "blockListUpdateInterval",
+			Usage: "Update list --blockDomainList,--blockCIDR4List,--blockCIDR6List interval, second. default 0, only read one time on start",
+		},
+		&cli.StringFlag{
+			Name:  "pid",
+			Usage: "A file path used to store pid. Send SIGUSR1 to me to reset the --serverLog file on unix system",
+		},
 	}
 	app.Before = func(c *cli.Context) error {
+		brook.ClientHKDFInfo = []byte(c.String("clientHKDFInfo"))
+		brook.ServerHKDFInfo = []byte(c.String("serverHKDFInfo"))
+		if c.String("pid") != "" {
+			if !filepath.IsAbs(c.String("pid")) {
+				return errors.New("--pid must be with absolute path")
+			}
+			if err := os.WriteFile(c.String("pid"), []byte(strconv.Itoa(os.Getpid())), 0744); err != nil {
+				return err
+			}
+		}
 		if c.String("pprof") != "" {
 			p, err := pprof.NewPprof(c.String("pprof"))
 			if err != nil {
@@ -82,12 +186,117 @@ func main() {
 				},
 			})
 		}
+		if c.String("log") != "" {
+			if c.String("log") != "console" && !filepath.IsAbs(c.String("log")) {
+				return errors.New("--log must be with absolute path")
+			}
+			var m map[string]string
+			if len(c.StringSlice("tag")) > 0 {
+				m = make(map[string]string)
+				for _, v := range c.StringSlice("tag") {
+					l := strings.Split(v, ":")
+					if len(l) != 2 {
+						return errors.New("Invalid tag " + v)
+					}
+					m[l[0]] = l[1]
+				}
+			}
+			p, err := logger.NewLogger(m, c.String("log"))
+			if err != nil {
+				return err
+			}
+			p.TouchBrook()
+			f := df
+			df = func() {
+				p.Close()
+				f()
+			}
+		}
+		if c.String("dialWithDNS") != "" {
+			p, err := dialwithdns.NewDialWithDNS(c.String("dialWithDNS"), c.String("dialWithDNSPrefer"))
+			if err != nil {
+				return err
+			}
+			p.TouchBrook()
+		}
+		if c.String("dialWithIP4") != "" || c.String("dialWithIP6") != "" {
+			p, err := dialwithip.NewDialWithIP(c.String("dialWithIP4"), c.String("dialWithIP6"))
+			if err != nil {
+				return err
+			}
+			p.TouchBrook()
+		}
+		if c.String("dialWithNIC") != "" {
+			p := dialwithnic.NewDialWithNIC(c.String("dialWithNIC"))
+			p.TouchBrook()
+		}
+		if c.String("dialWithSocks5") != "" {
+			p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("dialWithSocks5TCPTimeout"), c.Int("dialWithSocks5UDPTimeout"))
+			if err != nil {
+				return err
+			}
+			p.TouchBrook()
+		}
+		if c.String("prometheus") != "" {
+			if c.String("prometheusPath") == "" {
+				return errors.New("You forgot the --prometheusPath")
+			}
+			var m map[string]string
+			if len(c.StringSlice("tag")) > 0 {
+				m = make(map[string]string)
+				for _, v := range c.StringSlice("tag") {
+					l := strings.Split(v, ":")
+					if len(l) != 2 {
+						return errors.New("Invalid tag " + v)
+					}
+					m[l[0]] = l[1]
+				}
+			}
+			p := prometheus.NewPrometheus(c.String("prometheus"), c.String("prometheusPath"), m)
+			p.TouchBrook()
+			g.Add(&runnergroup.Runner{
+				Start: func() error {
+					return p.ListenAndServe()
+				},
+				Stop: func() error {
+					return p.Shutdown()
+				},
+			})
+		}
+		if c.String("blockDomainList") != "" && !strings.HasPrefix(c.String("blockDomainList"), "http://") && !strings.HasPrefix(c.String("blockDomainList"), "https://") && !filepath.IsAbs(c.String("blockDomainList")) {
+			return errors.New("--blockDomainList must be with absolute path")
+		}
+		if c.String("blockCIDR4List") != "" && !strings.HasPrefix(c.String("blockCIDR4List"), "http://") && !strings.HasPrefix(c.String("blockCIDR4List"), "https://") && !filepath.IsAbs(c.String("blockCIDR4List")) {
+			return errors.New("--blockCIDR4List must be with absolute path")
+		}
+		if c.String("blockCIDR6List") != "" && !strings.HasPrefix(c.String("blockCIDR6List"), "http://") && !strings.HasPrefix(c.String("blockCIDR6List"), "https://") && !filepath.IsAbs(c.String("blockCIDR6List")) {
+			return errors.New("--blockCIDR6List must be with absolute path")
+		}
+		if c.String("blockDomainList") != "" || c.String("blockCIDR4List") != "" || c.String("blockCIDR6List") != "" || len(c.StringSlice("blockGeoIP")) != 0 {
+			p, err := block.NewBlock(c.String("blockDomainList"), c.String("blockCIDR4List"), c.String("blockCIDR6List"), c.StringSlice("blockGeoIP"), c.Int("blockListUpdateInterval"))
+			if err != nil {
+				return err
+			}
+			p.TouchBrook()
+			if c.Int("blockListUpdateInterval") != 0 {
+				g.Add(&runnergroup.Runner{
+					Start: func() error {
+						p.Update()
+						return nil
+					},
+					Stop: func() error {
+						p.Stop()
+						return nil
+					},
+				})
+			}
+		}
 		return nil
 	}
 	app.Commands = []*cli.Command{
 		&cli.Command{
 			Name:  "server",
-			Usage: "Run as brook server, both TCP and UDP",
+			Usage: "Start a brook server that supports tcp and udp",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -117,38 +326,34 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR4List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr4.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR6List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr6.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringSliceFlag{
 					Name:  "blockGeoIP",
-					Usage: "Block IP by Geo country code, such as US",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.Int64Flag{
 					Name:  "updateListInterval",
-					Usage: "Update list interval, second. default 0, only read one time on start",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
-				&cli.StringFlag{
-					Name:  "toSocks5",
-					Usage: "Forward to socks5 server, requires your socks5 supports standard socks5 TCP and UDP, such as 1.2.3.4:1080",
-				},
-				&cli.StringFlag{
-					Name:  "toSocks5Username",
-					Usage: "Forward to socks5 server, username",
-				},
-				&cli.StringFlag{
-					Name:  "toSocks5Password",
-					Usage: "Forward to socks5 server, password",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook server --listen :9999 --password hello")
+					return nil
+				}
 				if c.String("listen") == "" || c.String("password") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -184,13 +389,6 @@ func main() {
 				if err != nil {
 					return err
 				}
-				if c.String("toSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("toSocks5"), c.String("toSocks5Username"), c.String("toSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
 						return s.ListenAndServe()
@@ -204,7 +402,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "client",
-			Usage: "Run as brook client, both TCP and UDP, to start a socks5 proxy, [src <-> socks5 <-> $ brook client <-> $ brook server <-> dst]",
+			Usage: "Start a brook client that supports tcp and udp. It can open a socks5 proxy, [src <-> socks5 <-> $ brook client <-> $ brook server <-> dst]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -222,9 +420,9 @@ func main() {
 					Aliases: []string{"p"},
 					Usage:   "Brook server password",
 				},
-				&cli.BoolFlag{
-					Name:  "udpovertcp",
-					Usage: "UDP over TCP",
+				&cli.StringFlag{
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The wssserver and password parameters will be ignored",
 				},
 				&cli.StringFlag{
 					Name:  "socks5",
@@ -249,49 +447,59 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5",
-					Usage: "Dial with your socks5 proxy, such as 127.0.0.1:1081",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Username",
-					Usage: "Optional",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Password",
-					Usage: "Optional",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("socks5") == "" || c.String("server") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook client --server 1.2.3.4:9999 --password hello --socks5 127.0.0.1:1080")
+					return nil
+				}
+				if c.String("server") == "" && c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
+				}
+				var link = ""
+				if c.String("server") != "" {
+					v := url.Values{}
+					v.Set("password", c.String("password"))
+					link = brook.Link("server", c.String("server"), v)
+				}
+				if c.String("link") != "" {
+					link = c.String("link")
 				}
 				h, p, err := net.SplitHostPort(c.String("socks5"))
 				if err != nil {
 					return err
 				}
-				if h == "" && c.String("socks5ServerIP") == "" {
+				if c.String("socks5ServerIP") != "" {
+					h = c.String("socks5ServerIP")
+				}
+				if h == "" {
 					return errors.New("socks5 server requires a clear IP for UDP, only port is not enough. You may use loopback IP or lan IP or other, we can not decide for you")
 				}
-				if c.String("dialWithSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				var ip string
-				if h != "" {
-					ip = h
-				}
-				if c.String("socks5ServerIP") != "" {
-					ip = c.String("socks5ServerIP")
-				}
-				s, err := brook.NewClient(c.String("socks5"), ip, c.String("server"), c.String("password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				kind, _, v, err := brook.ParseLink(link)
 				if err != nil {
 					return err
 				}
-				s.UDPOverTCP = c.Bool("udpovertcp")
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
+				}
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewBrookLink(link)
+				if err != nil {
+					return err
+				}
+				err = s.PrepareSocks5Server(c.String("socks5"), h, c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if err != nil {
+					return err
+				}
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
 						return s.ListenAndServe()
@@ -301,31 +509,25 @@ func main() {
 					},
 				})
 				if c.String("http") != "" {
-					h, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(ip, p), "", "", c.Int("tcpTimeout"))
+					s, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(h, p), "", "", c.Int("tcpTimeout"))
 					if err != nil {
 						return err
 					}
 					g.Add(&runnergroup.Runner{
 						Start: func() error {
-							return h.ListenAndServe()
+							return s.ListenAndServe()
 						},
 						Stop: func() error {
-							return h.Shutdown()
+							return s.Shutdown()
 						},
 					})
 				}
-				go func() {
-					sigs := make(chan os.Signal, 1)
-					signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
-					<-sigs
-					g.Done()
-				}()
-				return g.Wait()
+				return nil
 			},
 		},
 		&cli.Command{
 			Name:  "wsserver",
-			Usage: "Run as brook wsserver, both TCP and UDP, it will start a standard http server and websocket server",
+			Usage: "Start a brook wsserver that supports tcp and udp. It opens a standard http server and a websocket server",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -364,38 +566,38 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR4List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr4.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR6List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr6.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringSliceFlag{
 					Name:  "blockGeoIP",
-					Usage: "Block IP by Geo country code, such as US",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.Int64Flag{
 					Name:  "updateListInterval",
-					Usage: "Update list interval, second. default 0, only read one time on start",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
-				&cli.StringFlag{
-					Name:  "toSocks5",
-					Usage: "Forward to socks5 server, requires your socks5 supports standard socks5 TCP and UDP, such as 1.2.3.4:1080",
+				&cli.BoolFlag{
+					Name:  "xForwardedFor",
+					Usage: "Replace the from field in --log, note that this may be forged",
 				},
-				&cli.StringFlag{
-					Name:  "toSocks5Username",
-					Usage: "Forward to socks5 server, username",
-				},
-				&cli.StringFlag{
-					Name:  "toSocks5Password",
-					Usage: "Forward to socks5 server, password",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook wsserver --listen :9999 --password hello")
+					return nil
+				}
 				if c.String("listen") == "" || c.String("password") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -427,17 +629,11 @@ func main() {
 						})
 					}
 				}
-				if c.String("toSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("toSocks5"), c.String("toSocks5Username"), c.String("toSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
 				s, err := brook.NewWSServer(c.String("listen"), c.String("password"), "", c.String("path"), c.Int("tcpTimeout"), c.Int("udpTimeout"), c.Bool("withoutBrookProtocol"))
 				if err != nil {
 					return err
 				}
+				s.XForwardedFor = c.Bool("xForwardedFor")
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
 						return s.ListenAndServe()
@@ -451,7 +647,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "wsclient",
-			Usage: "Run as brook wsclient, both TCP and UDP, to start a socks5 proxy, [src <-> socks5 <-> $ brook wsclient <-> $ brook wsserver <-> dst]",
+			Usage: "Start a brook wsclient that supports tcp and udp. It can open a socks5 proxy, [src <-> socks5 <-> $ brook wsclient <-> $ brook wsserver <-> dst]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -470,12 +666,8 @@ func main() {
 					Usage:   "Brook wsserver password",
 				},
 				&cli.StringFlag{
-					Name:  "address",
-					Usage: "Specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "The data will not be encrypted with brook protocol",
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The wssserver and password parameters will be ignored",
 				},
 				&cli.StringFlag{
 					Name:  "socks5",
@@ -500,50 +692,58 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5",
-					Usage: "Dial with your socks5 proxy, such as 127.0.0.1:1081",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Username",
-					Usage: "Optional",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Password",
-					Usage: "Optional",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("socks5") == "" || c.String("wsserver") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook wsclient --wsserver ws://1.2.3.4:9999 --password hello --socks5 127.0.0.1:1080")
+					return nil
+				}
+				if c.String("wsserver") == "" && c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
+				}
+				var link = ""
+				if c.String("wsserver") != "" {
+					v := url.Values{}
+					v.Set("password", c.String("password"))
+					link = brook.Link("wsserver", c.String("wsserver"), v)
+				}
+				if c.String("link") != "" {
+					link = c.String("link")
 				}
 				h, p, err := net.SplitHostPort(c.String("socks5"))
 				if err != nil {
 					return err
 				}
-				if h == "" && c.String("socks5ServerIP") == "" {
+				if c.String("socks5ServerIP") != "" {
+					h = c.String("socks5ServerIP")
+				}
+				if h == "" {
 					return errors.New("socks5 server requires a clear IP for UDP, only port is not enough. You may use loopback IP or lan IP or other, we can not decide for you")
 				}
-				if c.String("dialWithSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				var ip string
-				if h != "" {
-					ip = h
-				}
-				if c.String("socks5ServerIP") != "" {
-					ip = c.String("socks5ServerIP")
-				}
-				s, err := brook.NewWSClient(c.String("socks5"), ip, c.String("wsserver"), c.String("password"), c.Int("tcpTimeout"), c.Int("udpTimeout"), c.Bool("withoutBrookProtocol"))
+				kind, _, v, err := brook.ParseLink(link)
 				if err != nil {
 					return err
 				}
-				if c.String("address") != "" {
-					s.ServerAddress = c.String("address")
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
+				}
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewBrookLink(link)
+				if err != nil {
+					return err
+				}
+				err = s.PrepareSocks5Server(c.String("socks5"), h, c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if err != nil {
+					return err
 				}
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
@@ -554,16 +754,16 @@ func main() {
 					},
 				})
 				if c.String("http") != "" {
-					h, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(ip, p), "", "", c.Int("tcpTimeout"))
+					s, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(h, p), "", "", c.Int("tcpTimeout"))
 					if err != nil {
 						return err
 					}
 					g.Add(&runnergroup.Runner{
 						Start: func() error {
-							return h.ListenAndServe()
+							return s.ListenAndServe()
 						},
 						Stop: func() error {
-							return h.Shutdown()
+							return s.Shutdown()
 						},
 					})
 				}
@@ -572,7 +772,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "wssserver",
-			Usage: "Run as brook wssserver, both TCP and UDP, it will start a standard https server and websocket server",
+			Usage: "Start a brook wssserver that supports tcp and udp. It opens a standard https server and a websocket server",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -618,38 +818,34 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR4List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr4.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR6List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr6.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringSliceFlag{
 					Name:  "blockGeoIP",
-					Usage: "Block IP by Geo country code, such as US",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.Int64Flag{
 					Name:  "updateListInterval",
-					Usage: "Update list interval, second. default 0, only read one time on start",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
-				&cli.StringFlag{
-					Name:  "toSocks5",
-					Usage: "Forward to socks5 server, requires your socks5 supports standard socks5 TCP and UDP, such as 1.2.3.4:1080",
-				},
-				&cli.StringFlag{
-					Name:  "toSocks5Username",
-					Usage: "Forward to socks5 server, username",
-				},
-				&cli.StringFlag{
-					Name:  "toSocks5Password",
-					Usage: "Forward to socks5 server, password",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook wssserver --domainaddress domain.com:9999 --password hello")
+					return nil
+				}
 				if c.String("domainaddress") == "" || c.String("password") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -687,13 +883,6 @@ func main() {
 						})
 					}
 				}
-				if c.String("toSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("toSocks5"), c.String("toSocks5Username"), c.String("toSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
 				h, p, err := net.SplitHostPort(c.String("domainaddress"))
 				if err != nil {
 					return err
@@ -703,14 +892,14 @@ func main() {
 					return err
 				}
 				if c.String("cert") != "" {
-					b, err := ioutil.ReadFile(c.String("cert"))
+					b, err := os.ReadFile(c.String("cert"))
 					if err != nil {
 						return err
 					}
 					s.Cert = b
 				}
 				if c.String("certkey") != "" {
-					b, err := ioutil.ReadFile(c.String("certkey"))
+					b, err := os.ReadFile(c.String("certkey"))
 					if err != nil {
 						return err
 					}
@@ -729,7 +918,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "wssclient",
-			Usage: "Run as brook wssclient, both TCP and UDP, to start a socks5 proxy, [src <-> socks5 <-> $ brook wssclient <-> $ brook wssserver <-> dst]",
+			Usage: "Start a brook wssclient that supports tcp and udp. It can open a socks5 proxy, [src <-> socks5 <-> $ brook wssclient <-> $ brook wssserver <-> dst]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -748,20 +937,8 @@ func main() {
 					Usage:   "Brook wssserver password",
 				},
 				&cli.StringFlag{
-					Name:  "address",
-					Usage: "Specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.BoolFlag{
-					Name:  "insecure",
-					Usage: "Client do not verify the server's certificate chain and host name",
-				},
-				&cli.StringFlag{
-					Name:  "ca",
-					Usage: "When server is brook wssserver, specify ca instead of insecure, such as /path/to/ca.pem",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "The data will not be encrypted with brook protocol",
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The wssserver and password parameters will be ignored",
 				},
 				&cli.StringFlag{
 					Name:  "socks5",
@@ -786,65 +963,58 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5",
-					Usage: "Dial with your socks5 proxy, such as 127.0.0.1:1081",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Username",
-					Usage: "Optional",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Password",
-					Usage: "Optional",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("socks5") == "" || c.String("wssserver") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook wssclient --wssserver wss://domain.com:9999 --password hello --socks5 127.0.0.1:1080")
+					return nil
+				}
+				if c.String("wssserver") == "" && c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
+				}
+				var link = ""
+				if c.String("wssserver") != "" {
+					v := url.Values{}
+					v.Set("password", c.String("password"))
+					link = brook.Link("wssserver", c.String("wssserver"), v)
+				}
+				if c.String("link") != "" {
+					link = c.String("link")
 				}
 				h, p, err := net.SplitHostPort(c.String("socks5"))
 				if err != nil {
 					return err
 				}
-				if h == "" && c.String("socks5ServerIP") == "" {
+				if c.String("socks5ServerIP") != "" {
+					h = c.String("socks5ServerIP")
+				}
+				if h == "" {
 					return errors.New("socks5 server requires a clear IP for UDP, only port is not enough. You may use loopback IP or lan IP or other, we can not decide for you")
 				}
-				if c.String("dialWithSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				var ip string
-				if h != "" {
-					ip = h
-				}
-				if c.String("socks5ServerIP") != "" {
-					ip = c.String("socks5ServerIP")
-				}
-				s, err := brook.NewWSClient(c.String("socks5"), ip, c.String("wssserver"), c.String("password"), c.Int("tcpTimeout"), c.Int("udpTimeout"), c.Bool("withoutBrookProtocol"))
+				kind, _, v, err := brook.ParseLink(link)
 				if err != nil {
 					return err
 				}
-				if c.String("address") != "" {
-					s.ServerAddress = c.String("address")
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
 				}
-				if c.Bool("insecure") {
-					s.TLSConfig.InsecureSkipVerify = true
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
 				}
-				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
-					if err != nil {
-						return err
-					}
-					roots := x509.NewCertPool()
-					ok := roots.AppendCertsFromPEM(b)
-					if !ok {
-						return errors.New("failed to parse root certificate")
-					}
-					s.TLSConfig.RootCAs = roots
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewBrookLink(link)
+				if err != nil {
+					return err
+				}
+				err = s.PrepareSocks5Server(c.String("socks5"), h, c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if err != nil {
+					return err
 				}
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
@@ -855,16 +1025,16 @@ func main() {
 					},
 				})
 				if c.String("http") != "" {
-					h, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(ip, p), "", "", c.Int("tcpTimeout"))
+					s, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(h, p), "", "", c.Int("tcpTimeout"))
 					if err != nil {
 						return err
 					}
 					g.Add(&runnergroup.Runner{
 						Start: func() error {
-							return h.ListenAndServe()
+							return s.ListenAndServe()
 						},
 						Stop: func() error {
-							return h.Shutdown()
+							return s.Shutdown()
 						},
 					})
 				}
@@ -873,7 +1043,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "quicserver",
-			Usage: "Run as brook quicserver, both TCP and UDP",
+			Usage: "Start a brook quicserver that supports tcp and udp.",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -914,26 +1084,34 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR4List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr4.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringFlag{
 					Name:  "blockCIDR6List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr6.txt",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.StringSliceFlag{
 					Name:  "blockGeoIP",
-					Usage: "Block IP by Geo country code, such as US",
+					Usage: "This option will be removed in a future version, please use the global option instead",
 				},
 				&cli.Int64Flag{
 					Name:  "updateListInterval",
-					Usage: "Update list interval, second. default 0, only read one time on start",
+					Usage: "This option will be removed in a future version, please use the global option instead",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook quicserver --domainaddress domain.com:9999 --password hello")
+					return nil
+				}
 				if c.String("domainaddress") == "" || c.String("password") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -980,14 +1158,14 @@ func main() {
 					return err
 				}
 				if c.String("cert") != "" {
-					b, err := ioutil.ReadFile(c.String("cert"))
+					b, err := os.ReadFile(c.String("cert"))
 					if err != nil {
 						return err
 					}
 					s.Cert = b
 				}
 				if c.String("certkey") != "" {
-					b, err := ioutil.ReadFile(c.String("certkey"))
+					b, err := os.ReadFile(c.String("certkey"))
 					if err != nil {
 						return err
 					}
@@ -1006,7 +1184,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "quicclient",
-			Usage: "Run as brook quicclient, both TCP and UDP, to start a socks5 proxy, [src <-> socks5 <-> $ brook quicclient <-> $ brook quicserver <-> dst]",
+			Usage: "Start a brook quicclient that supports tcp and udp. It can open a socks5 proxy, [src <-> socks5 <-> $ brook quicclient <-> $ brook quicserver <-> dst]. (The global-dial-parameter is ignored)",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -1015,26 +1193,8 @@ func main() {
 			},
 			Flags: []cli.Flag{
 				&cli.StringFlag{
-					Name:    "quicserver",
-					Aliases: []string{"s"},
-					Usage:   "Brook quicserver address, like: quic://google.com:443. Do not omit the port under any circumstances",
-				},
-				&cli.StringFlag{
-					Name:    "password",
-					Aliases: []string{"p"},
-					Usage:   "Brook quicserver password",
-				},
-				&cli.StringFlag{
-					Name:  "address",
-					Usage: "Specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.StringFlag{
-					Name:  "ca",
-					Usage: "When server is brook wssserver, specify ca instead of insecure, such as /path/to/ca.pem",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "The data will not be encrypted with brook protocol",
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The wssserver and password parameters will be ignored",
 				},
 				&cli.StringFlag{
 					Name:  "socks5",
@@ -1059,43 +1219,50 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("socks5") == "" || c.String("quicserver") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook quicclient --link 'brook://...' --socks5 127.0.0.1:1080")
+					return nil
+				}
+				if c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
+				link := c.String("link")
 				h, p, err := net.SplitHostPort(c.String("socks5"))
 				if err != nil {
 					return err
 				}
-				if h == "" && c.String("socks5ServerIP") == "" {
+				if c.String("socks5ServerIP") != "" {
+					h = c.String("socks5ServerIP")
+				}
+				if h == "" {
 					return errors.New("socks5 server requires a clear IP for UDP, only port is not enough. You may use loopback IP or lan IP or other, we can not decide for you")
 				}
-				var ip string
-				if h != "" {
-					ip = h
-				}
-				if c.String("socks5ServerIP") != "" {
-					ip = c.String("socks5ServerIP")
-				}
-				s, err := brook.NewQUICClient(c.String("socks5"), ip, c.String("quicserver"), c.String("password"), c.Int("tcpTimeout"), c.Int("udpTimeout"), c.Bool("withoutBrookProtocol"))
+				kind, _, v, err := brook.ParseLink(link)
 				if err != nil {
 					return err
 				}
-				if c.String("address") != "" {
-					s.ServerAddress = c.String("address")
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
 				}
-				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
-					if err != nil {
-						return err
-					}
-					roots := x509.NewCertPool()
-					ok := roots.AppendCertsFromPEM(b)
-					if !ok {
-						return errors.New("failed to parse root certificate")
-					}
-					s.TLSConfig.RootCAs = roots
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewBrookLink(link)
+				if err != nil {
+					return err
+				}
+				err = s.PrepareSocks5Server(c.String("socks5"), h, c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if err != nil {
+					return err
 				}
 				g.Add(&runnergroup.Runner{
 					Start: func() error {
@@ -1106,16 +1273,16 @@ func main() {
 					},
 				})
 				if c.String("http") != "" {
-					h, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(ip, p), "", "", c.Int("tcpTimeout"))
+					s, err := brook.NewSocks5ToHTTP(c.String("http"), net.JoinHostPort(h, p), "", "", c.Int("tcpTimeout"))
 					if err != nil {
 						return err
 					}
 					g.Add(&runnergroup.Runner{
 						Start: func() error {
-							return h.ListenAndServe()
+							return s.ListenAndServe()
 						},
 						Stop: func() error {
-							return h.Shutdown()
+							return s.Shutdown()
 						},
 					})
 				}
@@ -1124,7 +1291,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "relayoverbrook",
-			Usage: "Run as relay over brook, both TCP and UDP, this means access [from address] is equal to [to address], [src <-> from address <-> $ brook server/wsserver/wssserver/quicserver <-> to address]",
+			Usage: "Relay network traffic over brook, which supports TCP and UDP. Accessing [from address] is equal to accessing [to address], [src <-> from address <-> $ brook server/wsserver/wssserver/quicserver <-> to address]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -1132,6 +1299,20 @@ func main() {
 				}
 			},
 			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:    "server",
+					Aliases: []string{"s"},
+					Usage:   "brook server or brook wsserver or brook wssserver or brook quicserver, like: 1.2.3.4:9999, ws://1.2.3.4:9999, wss://domain:443/ws, quic://domain.com:443",
+				},
+				&cli.StringFlag{
+					Name:    "password",
+					Aliases: []string{"p"},
+					Usage:   "Password",
+				},
+				&cli.StringFlag{
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The server and password parameters will be ignored",
+				},
 				&cli.StringFlag{
 					Name:    "from",
 					Aliases: []string{"f", "l"},
@@ -1141,36 +1322,6 @@ func main() {
 					Name:    "to",
 					Aliases: []string{"t"},
 					Usage:   "Address which relay to, like: 1.2.3.4:9999",
-				},
-				&cli.StringFlag{
-					Name:    "server",
-					Aliases: []string{"s"},
-					Usage:   "brook server or brook wsserver or brook wssserver, like: 1.2.3.4:9999, ws://1.2.3.4:9999, wss://domain:443/ws, quic://domain.com:443",
-				},
-				&cli.StringFlag{
-					Name:    "password",
-					Aliases: []string{"p"},
-					Usage:   "Password",
-				},
-				&cli.BoolFlag{
-					Name:  "udpovertcp",
-					Usage: "When server is brook server, UDP over TCP",
-				},
-				&cli.StringFlag{
-					Name:  "address",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.BoolFlag{
-					Name:  "insecure",
-					Usage: "When server is brook wssserver or brook quicserver, client do not verify the server's certificate chain and host name",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, the data will not be encrypted with brook protocol",
-				},
-				&cli.StringFlag{
-					Name:  "ca",
-					Usage: "When server is brook wssserver or brook quicserver, specify ca instead of insecure, such as /path/to/ca.pem",
 				},
 				&cli.IntFlag{
 					Name:  "tcpTimeout",
@@ -1182,43 +1333,55 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("from") == "" || c.String("to") == "" || c.String("server") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook relayoverbrook --server 1.2.3.4:9999 --password hello --from :5353 --to 8.8.8.8:53")
+					return nil
+				}
+				if c.String("from") == "" || c.String("to") == "" || (c.String("server") == "" && c.String("link") == "") {
 					return cli.ShowSubcommandHelp(c)
 				}
-				kind := "server"
-				if strings.HasPrefix(c.String("server"), "ws://") {
-					kind = "wsserver"
-				}
-				if strings.HasPrefix(c.String("server"), "wss://") {
-					kind = "wssserver"
-				}
-				if strings.HasPrefix(c.String("server"), "quic://") {
-					kind = "quicserver"
-				}
-				v := url.Values{}
-				if c.Bool("udpovertcp") {
-					v.Set("udpovertcp", "true")
-				}
-				if c.String("address") != "" {
-					v.Set("address", c.String("address"))
-				}
-				if c.Bool("insecure") {
-					v.Set("insecure", "true")
-				}
-				if c.Bool("withoutBrookProtocol") {
-					v.Set("withoutBrookProtocol", "true")
-				}
-				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
-					if err != nil {
-						return err
+				var link = ""
+				if c.String("server") != "" {
+					kind := "server"
+					if strings.HasPrefix(c.String("server"), "ws://") {
+						kind = "wsserver"
 					}
-					v.Set("ca", string(b))
+					if strings.HasPrefix(c.String("server"), "wss://") {
+						kind = "wssserver"
+					}
+					if strings.HasPrefix(c.String("server"), "quic://") {
+						kind = "quicserver"
+					}
+					if kind == "quicserver" {
+						return errors.New("It is recommended to use brook link and specify --udpoverstream")
+					}
+					v := url.Values{}
+					v.Set("password", c.String("password"))
+					link = brook.Link(kind, c.String("server"), v)
 				}
-				v.Set("password", c.String("password"))
-				s, err := brook.NewRelayOverBrook(c.String("from"), brook.Link(kind, c.String("server"), v), c.String("to"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if c.String("link") != "" {
+					link = c.String("link")
+				}
+				kind, _, v, err := brook.ParseLink(link)
+				if err != nil {
+					return err
+				}
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
+				}
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewRelayOverBrook(c.String("from"), link, c.String("to"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
 				if err != nil {
 					return err
 				}
@@ -1235,7 +1398,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "dnsserveroverbrook",
-			Usage: "Run as dns server over brook, both TCP and UDP, [src <-> $ brook dnserversoverbrook <-> $ brook server/wsserver/wssserver/quicserver <-> dns] or [src <-> $ brook dnsserveroverbrook <-> dnsForBypass]",
+			Usage: "Run a dns server over brook, which supports TCP and UDP, [src <-> $ brook dnserversoverbrook <-> $ brook server/wsserver/wssserver/quicserver <-> dns] or [src <-> $ brook dnsserveroverbrook <-> dnsForBypass]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -1243,6 +1406,20 @@ func main() {
 				}
 			},
 			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:    "server",
+					Aliases: []string{"s"},
+					Usage:   "brook server or brook wsserver or brook wssserver or brook quicserver, like: 1.2.3.4:9999, ws://1.2.3.4:9999, wss://domain.com:443/ws, quic://domain.com:443",
+				},
+				&cli.StringFlag{
+					Name:    "password",
+					Aliases: []string{"p"},
+					Usage:   "Password",
+				},
+				&cli.StringFlag{
+					Name:  "link",
+					Usage: "brook link, you can get it via $ brook link. The server and password parameters will be ignored",
+				},
 				&cli.StringFlag{
 					Name:    "listen",
 					Aliases: []string{"l"},
@@ -1255,16 +1432,16 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "dnsForBypass",
-					Usage: "DNS server for resolving domains in bypass list",
+					Usage: "DNS server for resolving domains in bypass list. Such as 223.5.5.5:53 or https://dns.alidns.com/dns-query?address=223.5.5.5:443, the address is required",
 					Value: "223.5.5.5:53",
 				},
 				&cli.StringFlag{
 					Name:  "bypassDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt",
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt",
 				},
 				&cli.BoolFlag{
 					Name:  "disableA",
@@ -1273,36 +1450,6 @@ func main() {
 				&cli.BoolFlag{
 					Name:  "disableAAAA",
 					Usage: "Disable AAAA query",
-				},
-				&cli.StringFlag{
-					Name:    "server",
-					Aliases: []string{"s"},
-					Usage:   "brook server or brook wsserver or brook wssserver, like: 1.2.3.4:9999, ws://1.2.3.4:9999, wss://domain.com:443/ws, quic://domain.com:443",
-				},
-				&cli.StringFlag{
-					Name:    "password",
-					Aliases: []string{"p"},
-					Usage:   "Password",
-				},
-				&cli.BoolFlag{
-					Name:  "udpovertcp",
-					Usage: "When server is brook server, UDP over TCP",
-				},
-				&cli.StringFlag{
-					Name:  "address",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.BoolFlag{
-					Name:  "insecure",
-					Usage: "When server is brook wssserver or brook quicserver, client do not verify the server's certificate chain and host name",
-				},
-				&cli.StringFlag{
-					Name:  "ca",
-					Usage: "When server is brook wssserver or brook quicserver, specify ca instead of insecure, such as /path/to/ca.pem",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, the data will not be encrypted with brook protocol",
 				},
 				&cli.IntFlag{
 					Name:  "tcpTimeout",
@@ -1314,9 +1461,17 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
-				if c.String("listen") == "" || c.String("server") == "" || c.String("password") == "" {
+				if c.Bool("example") {
+					fmt.Println("brook dnsserveroverbrook --server 1.2.3.4:9999 --password hello --listen 127.0.0.1:53")
+					return nil
+				}
+				if c.String("listen") == "" || (c.String("server") == "" && c.String("link") == "") {
 					return cli.ShowSubcommandHelp(c)
 				}
 				if c.String("bypassDomainList") != "" && !strings.HasPrefix(c.String("bypassDomainList"), "http://") && !strings.HasPrefix(c.String("bypassDomainList"), "https://") && !filepath.IsAbs(c.String("bypassDomainList")) {
@@ -1326,44 +1481,48 @@ func main() {
 					return errors.New("--blockDomainList must be with absolute path")
 				}
 				if c.String("blockDomainList") != "" || c.String("bypassDomainList") != "" || c.Bool("disableA") || c.Bool("disableAAAA") {
-					p, err := thedns.NewTheDNS(c.String("blockDomainList"), c.String("bypassDomainList"), c.String("dnsForBypass"), c.Bool("disableA"), c.Bool("disableAAAA"))
+					p, err := thedns.NewTheDNS(c.String("blockDomainList"), c.String("bypassDomainList"), c.String("dnsForBypass"), c.Bool("disableA"), c.Bool("disableAAAA"), "")
 					if err != nil {
 						return err
 					}
 					p.TouchBrook()
 				}
-				kind := "server"
-				if strings.HasPrefix(c.String("server"), "ws://") {
-					kind = "wsserver"
-				}
-				if strings.HasPrefix(c.String("server"), "wss://") {
-					kind = "wssserver"
-				}
-				if strings.HasPrefix(c.String("server"), "quic://") {
-					kind = "quicserver"
-				}
-				v := url.Values{}
-				if c.Bool("udpovertcp") {
-					v.Set("udpovertcp", "true")
-				}
-				if c.String("address") != "" {
-					v.Set("address", c.String("address"))
-				}
-				if c.Bool("insecure") {
-					v.Set("insecure", "true")
-				}
-				if c.Bool("withoutBrookProtocol") {
-					v.Set("withoutBrookProtocol", "true")
-				}
-				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
-					if err != nil {
-						return err
+				var link = ""
+				if c.String("server") != "" {
+					kind := "server"
+					if strings.HasPrefix(c.String("server"), "ws://") {
+						kind = "wsserver"
 					}
-					v.Set("ca", string(b))
+					if strings.HasPrefix(c.String("server"), "wss://") {
+						kind = "wssserver"
+					}
+					if strings.HasPrefix(c.String("server"), "quic://") {
+						kind = "quicserver"
+					}
+					if kind == "quicserver" {
+						return errors.New("It is recommended to use brook link and specify --udpoverstream")
+					}
+					v := url.Values{}
+					v.Set("password", c.String("password"))
+					link = brook.Link(kind, c.String("server"), v)
 				}
-				v.Set("password", c.String("password"))
-				s, err := brook.NewRelayOverBrook(c.String("listen"), brook.Link(kind, c.String("server"), v), c.String("dns"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if c.String("link") != "" {
+					link = c.String("link")
+				}
+				kind, _, v, err := brook.ParseLink(link)
+				if err != nil {
+					return err
+				}
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
+				}
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
+				if kind == "socks5" {
+					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
+				}
+				s, err := brook.NewRelayOverBrook(c.String("listen"), link, c.String("dns"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
 				if err != nil {
 					return err
 				}
@@ -1380,399 +1539,8 @@ func main() {
 			},
 		},
 		&cli.Command{
-			Name:  "tproxy",
-			Usage: "Run as transparent proxy, both TCP and UDP, only works on Linux, [src <-> $ brook tproxy <-> $ brook server/wsserver/wssserver/quicserver <-> dst]",
-			BashComplete: func(c *cli.Context) {
-				l := c.Command.VisibleFlags()
-				for _, v := range l {
-					fmt.Println("--" + v.Names()[0])
-				}
-			},
-			Flags: []cli.Flag{
-				&cli.StringFlag{
-					Name:    "listen",
-					Aliases: []string{"l"},
-					Usage:   "Listen address, DO NOT contain IP, just like: ':1080'. No need to operate iptables by default!",
-					Value:   ":1080",
-				},
-				&cli.StringFlag{
-					Name:  "dnsListen",
-					Usage: "Start a smart DNS server, like: ':53'",
-				},
-				&cli.StringFlag{
-					Name:  "dnsForDefault",
-					Usage: "DNS server for resolving domains NOT in list",
-					Value: "8.8.8.8:53",
-				},
-				&cli.StringFlag{
-					Name:  "dnsForBypass",
-					Usage: "DNS server for resolving domains in bypass list",
-					Value: "223.5.5.5:53",
-				},
-				&cli.StringFlag{
-					Name:  "blockDomainList",
-					Usage: "One domain per line, Suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
-				},
-				&cli.StringFlag{
-					Name:  "bypassDomainList",
-					Usage: "One domain per line, Suffix match mode. https://, http:// or local file absolute path. Like: https://txthinking.github.io/bypass/example_domain.txt",
-				},
-				&cli.BoolFlag{
-					Name:  "disableA",
-					Usage: "Disable A query",
-				},
-				&cli.BoolFlag{
-					Name:  "disableAAAA",
-					Usage: "Disable AAAA query",
-				},
-				&cli.StringFlag{
-					Name:  "bypassCIDR4List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr4.txt",
-				},
-				&cli.StringFlag{
-					Name:  "bypassCIDR6List",
-					Usage: "One CIDR per line, https://, http:// or local file absolute path, like: https://txthinking.github.io/bypass/example_cidr6.txt",
-				},
-				&cli.StringSliceFlag{
-					Name:  "bypassGeoIP",
-					Usage: "Bypass IP by Geo country code, such as US",
-				},
-				&cli.BoolFlag{
-					Name:  "enableIPv6",
-					Usage: "deprecated",
-				},
-				&cli.BoolFlag{
-					Name:  "doNotRunScripts",
-					Usage: "This will not change iptables and others if you want to do by yourself",
-				},
-				&cli.StringFlag{
-					Name:  "webListen",
-					Usage: "Ignore all other parameters, run web UI, like: ':9999'",
-				},
-				&cli.StringFlag{
-					Name:    "server",
-					Aliases: []string{"s"},
-					Usage:   "brook server or brook wsserver or brook wssserver, like: 1.2.3.4:9999, ws://1.2.3.4:9999, wss://domain.com:443/ws, quic://domain.com:443",
-				},
-				&cli.StringFlag{
-					Name:    "password",
-					Aliases: []string{"p"},
-					Usage:   "Password",
-				},
-				&cli.BoolFlag{
-					Name:  "udpovertcp",
-					Usage: "When server is brook server, UDP over TCP",
-				},
-				&cli.StringFlag{
-					Name:  "address",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, specify address instead of resolving addresses from host, such as 1.2.3.4:443",
-				},
-				&cli.BoolFlag{
-					Name:  "insecure",
-					Usage: "When server is brook wssserver or brook quicserver, client do not verify the server's certificate chain and host name",
-				},
-				&cli.BoolFlag{
-					Name:  "withoutBrookProtocol",
-					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, the data will not be encrypted with brook protocol",
-				},
-				&cli.StringFlag{
-					Name:  "ca",
-					Usage: "When server is brook wssserver or brook quicserver, specify ca instead of insecure, such as /path/to/ca.pem",
-				},
-				&cli.StringFlag{
-					Name:  "link",
-					Usage: "brook link. This will ignore server, password, udpovertcp, address, insecure, withoutBrookProtocol, ca",
-				},
-				&cli.IntFlag{
-					Name:  "tcpTimeout",
-					Value: 0,
-					Usage: "time (s)",
-				},
-				&cli.IntFlag{
-					Name:  "udpTimeout",
-					Value: 60,
-					Usage: "time (s)",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5",
-					Usage: "Dial with your socks5 proxy, such as 127.0.0.1:1081",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Username",
-					Usage: "Optional",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Password",
-					Usage: "Optional",
-				},
-			},
-			Action: func(c *cli.Context) error {
-				if c.String("webListen") != "" {
-					web, err := fs.Sub(static, "static")
-					if err != nil {
-						return err
-					}
-					var cmd *exec.Cmd
-					lock := &sync.Mutex{}
-					m := http.NewServeMux()
-					m.Handle("/", http.FileServer(http.FS(web)))
-					m.HandleFunc("/hasp", func(w http.ResponseWriter, r *http.Request) {
-						lock.Lock()
-						defer lock.Unlock()
-						_, err := os.Stat("/root/.brook.web.password")
-						if os.IsNotExist(err) {
-							w.Write([]byte("no"))
-							return
-						}
-						w.Write([]byte("yes"))
-					})
-					m.HandleFunc("/setp", func(w http.ResponseWriter, r *http.Request) {
-						lock.Lock()
-						defer lock.Unlock()
-						_, err := os.Stat("/root/.brook.web.password")
-						if !os.IsNotExist(err) {
-							http.Error(w, "file exsits", 500)
-							return
-						}
-						err = ioutil.WriteFile("/root/.brook.web.password", []byte(r.FormValue("p")), 0600)
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						w.WriteHeader(200)
-					})
-					m.HandleFunc("/authp", func(w http.ResponseWriter, r *http.Request) {
-						lock.Lock()
-						defer lock.Unlock()
-						b, err := ioutil.ReadFile("/root/.brook.web.password")
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						if string(b) != r.FormValue("p") {
-							http.Error(w, "web ui password wrong", 500)
-							return
-						}
-						w.WriteHeader(200)
-					})
-					m.HandleFunc("/start", func(w http.ResponseWriter, r *http.Request) {
-						b, err := ioutil.ReadFile("/root/.brook.web.password")
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						if string(b) != r.FormValue("p") {
-							http.Error(w, "web ui password wrong", 500)
-							return
-						}
-						s, err := os.Executable()
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						lock.Lock()
-						defer lock.Unlock()
-						cmd = exec.Command("/bin/sh", "-c", s+" tproxy "+r.FormValue("args"))
-						done := make(chan byte)
-						defer close(done)
-						errch := make(chan error)
-						go func() {
-							out, _ := cmd.CombinedOutput()
-							select {
-							case <-done:
-								log.Println(string(out))
-							default:
-								select {
-								case <-done:
-									log.Println(string(out))
-								case errch <- errors.New(string(out)):
-								}
-							}
-							lock.Lock()
-							cmd = nil
-							lock.Unlock()
-						}()
-						select {
-						case err := <-errch:
-							http.Error(w, err.Error(), 500)
-							return
-						case <-time.After(6 * time.Second):
-							w.Write([]byte("connected"))
-						}
-					})
-					m.HandleFunc("/stop", func(w http.ResponseWriter, r *http.Request) {
-						b, err := ioutil.ReadFile("/root/.brook.web.password")
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						if string(b) != r.FormValue("p") {
-							http.Error(w, "web ui password wrong", 500)
-							return
-						}
-						lock.Lock()
-						defer lock.Unlock()
-						if cmd == nil {
-							w.Write([]byte("disconnected"))
-							return
-						}
-						if err := cmd.Process.Signal(syscall.SIGTERM); err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						w.Write([]byte("disconnected"))
-					})
-					m.HandleFunc("/status", func(w http.ResponseWriter, r *http.Request) {
-						b, err := ioutil.ReadFile("/root/.brook.web.password")
-						if err != nil {
-							http.Error(w, err.Error(), 500)
-							return
-						}
-						if string(b) != r.FormValue("p") {
-							http.Error(w, "web ui password wrong", 500)
-							return
-						}
-						lock.Lock()
-						defer lock.Unlock()
-						if cmd == nil {
-							w.Write([]byte("disconnected"))
-							return
-						}
-						w.Write([]byte("connected"))
-					})
-					s := &http.Server{
-						Addr:    c.String("webListen"),
-						Handler: m,
-					}
-					g.Add(&runnergroup.Runner{
-						Start: func() error {
-							return s.ListenAndServe()
-						},
-						Stop: func() error {
-							if cmd != nil {
-								cmd.Process.Signal(syscall.SIGTERM)
-							}
-							return s.Shutdown(context.Background())
-						},
-					})
-					return nil
-				}
-				if c.String("listen") == "" || (c.String("link") == "" && (c.String("server") == "" || c.String("password") == "")) {
-					_ = cli.ShowSubcommandHelp(c)
-					return errors.New("")
-				}
-				if c.String("blockDomainList") != "" && !strings.HasPrefix(c.String("blockDomainList"), "http://") && !strings.HasPrefix(c.String("blockDomainList"), "https://") && !filepath.IsAbs(c.String("blockDomainList")) {
-					return errors.New("--blockDomainList must be with absolute path")
-				}
-				if c.String("bypassDomainList") != "" && !strings.HasPrefix(c.String("bypassDomainList"), "http://") && !strings.HasPrefix(c.String("bypassDomainList"), "https://") && !filepath.IsAbs(c.String("bypassDomainList")) {
-					return errors.New("--bypassDomainList must be with absolute path")
-				}
-				if c.String("bypassCIDR4List") != "" && !strings.HasPrefix(c.String("bypassCIDR4List"), "http://") && !strings.HasPrefix(c.String("bypassCIDR4List"), "https://") && !filepath.IsAbs(c.String("bypassCIDR4List")) {
-					return errors.New("--bypassCIDR4List must be with absolute path")
-				}
-				if c.String("bypassCIDR6List") != "" && !strings.HasPrefix(c.String("bypassCIDR6List"), "http://") && !strings.HasPrefix(c.String("bypassCIDR6List"), "https://") && !filepath.IsAbs(c.String("bypassCIDR6List")) {
-					return errors.New("--bypassCIDR6List must be with absolute path")
-				}
-				if c.String("blockDomainList") != "" || c.String("bypassDomainList") != "" || c.Bool("disableA") || c.Bool("disableAAAA") {
-					p, err := thedns.NewTheDNS(c.String("blockDomainList"), c.String("bypassDomainList"), c.String("dnsForBypass"), c.Bool("disableA"), c.Bool("disableAAAA"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				if c.String("bypassCIDR4List") != "" || c.String("bypassCIDR6List") != "" || len(c.StringSlice("bypassGeoIP")) != 0 {
-					p, err := tproxy.NewTproxy(c.String("bypassCIDR4List"), c.String("bypassCIDR6List"), c.StringSlice("bypassGeoIP"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				if c.String("dialWithSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
-				}
-				kind := "server"
-				if strings.HasPrefix(c.String("server"), "ws://") {
-					kind = "wsserver"
-				}
-				if strings.HasPrefix(c.String("server"), "wss://") {
-					kind = "wssserver"
-				}
-				if strings.HasPrefix(c.String("server"), "quic://") {
-					kind = "quicserver"
-				}
-				v := url.Values{}
-				if c.Bool("udpovertcp") {
-					v.Set("udpovertcp", "true")
-				}
-				if c.String("address") != "" {
-					v.Set("address", c.String("address"))
-				}
-				if c.Bool("insecure") {
-					v.Set("insecure", "true")
-				}
-				if c.Bool("withoutBrookProtocol") {
-					v.Set("withoutBrookProtocol", "true")
-				}
-				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
-					if err != nil {
-						return err
-					}
-					v.Set("ca", string(b))
-				}
-				v.Set("password", c.String("password"))
-				link := brook.Link(kind, c.String("server"), v)
-				if c.String("link") != "" {
-					link = c.String("link")
-				}
-				s, err := brook.NewTproxy(c.String("listen"), link, c.Int("tcpTimeout"), c.Int("udpTimeout"))
-				if err != nil {
-					return err
-				}
-				g.Add(&runnergroup.Runner{
-					Start: func() error {
-						if !c.Bool("doNotRunScripts") {
-							s.ClearAutoScripts()
-							if err := s.RunAutoScripts(); err != nil {
-								return err
-							}
-						}
-						return s.ListenAndServe()
-					},
-					Stop: func() error {
-						if !c.Bool("doNotRunScripts") {
-							if err := s.ClearAutoScripts(); err != nil {
-								log.Println(err)
-							}
-						}
-						return s.Shutdown()
-					},
-				})
-				if c.String("dnsListen") != "" {
-					s, err := brook.NewRelayOverBrook(c.String("dnsListen"), link, c.String("dnsForDefault"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					s.IsDNS = true
-					g.Add(&runnergroup.Runner{
-						Start: func() error {
-							return s.ListenAndServe()
-						},
-						Stop: func() error {
-							return s.Shutdown()
-						},
-					})
-				}
-				return nil
-			},
-		},
-		&cli.Command{
 			Name:  "link",
-			Usage: "Generate brook link",
+			Usage: "Generate a brook link",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -1803,6 +1571,10 @@ func main() {
 					Name:  "udpovertcp",
 					Usage: "When server is brook server, UDP over TCP",
 				},
+				&cli.BoolFlag{
+					Name:  "udpoverstream",
+					Usage: "When server is brook quicserver, UDP over Stream. Under normal circumstances, you need this parameter because the max datagram size for QUIC is very small. Note: only brook CLI and tun2brook suppport for now",
+				},
 				&cli.StringFlag{
 					Name:  "address",
 					Usage: "When server is brook wsserver or brook wssserver or brook quicserver, specify address instead of resolving addresses from host, such as 1.2.3.4:443",
@@ -1819,8 +1591,36 @@ func main() {
 					Name:  "ca",
 					Usage: "When server is brook wssserver or brook quicserver, specify ca for untrusted cert, such as /path/to/ca.pem",
 				},
+				&cli.StringFlag{
+					Name:  "tlsfingerprint",
+					Usage: "When server is brook wssserver, select tls fingerprint, value can be: chrome",
+				},
+				&cli.StringFlag{
+					Name:  "clientHKDFInfo",
+					Usage: "client HKDF info, most time you don't need to change this, read brook protocol if you don't know what this is",
+				},
+				&cli.StringFlag{
+					Name:  "serverHKDFInfo",
+					Usage: "server HKDF info, most time you don't need to change this, read brook protocol if you don't know what this is",
+				},
+				&cli.StringFlag{
+					Name:  "fragment",
+					Usage: "When server is brook wssserver, split the ClientHello into multiple fragments and then send them one by one with delays (millisecond). The format is min_length:max_length:min_delay:max_delay, cannot be zero, such as 50:100:10:50",
+				},
+				&cli.StringFlag{
+					Name:  "token",
+					Usage: "A token represents a user's identity. A string encoded in hexadecimal. Server needs to have --userAPI enabled. Note that: Only supported by the brook GUI(except for OpenWrt) and tun2brook",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook link --server 1.2.3.4:9999 --password hello --name 'my server'")
+					return nil
+				}
 				if c.String("server") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -1848,6 +1648,9 @@ func main() {
 				if c.Bool("udpovertcp") {
 					v.Set("udpovertcp", "true")
 				}
+				if c.Bool("udpoverstream") {
+					v.Set("udpoverstream", "true")
+				}
 				if c.String("address") != "" {
 					v.Set("address", c.String("address"))
 				}
@@ -1857,12 +1660,27 @@ func main() {
 				if c.Bool("withoutBrookProtocol") {
 					v.Set("withoutBrookProtocol", "true")
 				}
+				if c.String("tlsfingerprint") != "" {
+					v.Set("tlsfingerprint", c.String("tlsfingerprint"))
+				}
 				if c.String("ca") != "" {
-					b, err := ioutil.ReadFile(c.String("ca"))
+					b, err := os.ReadFile(c.String("ca"))
 					if err != nil {
 						return err
 					}
 					v.Set("ca", string(b))
+				}
+				if c.String("clientHKDFInfo") != "" {
+					v.Set("clientHKDFInfo", c.String("clientHKDFInfo"))
+				}
+				if c.String("serverHKDFInfo") != "" {
+					v.Set("serverHKDFInfo", c.String("serverHKDFInfo"))
+				}
+				if c.String("fragment") != "" {
+					v.Set("fragment", c.String("fragment"))
+				}
+				if c.String("token") != "" {
+					v.Set("token", c.String("token"))
 				}
 				fmt.Println(brook.Link(s, c.String("server"), v))
 				return nil
@@ -1870,7 +1688,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "connect",
-			Usage: "Run as client and connect to brook link, both TCP and UDP, to start a socks5 proxy, [src <-> socks5 <-> $ brook connect <-> $ brook server/wsserver/wssserver/quicserver <-> dst]",
+			Usage: "Run a client and connect with a brook link, which supports TCP and UDP. It can start a socks5 proxy, [src <-> socks5 <-> $ brook connect <-> $ brook server/wsserver/wssserver/quicserver <-> dst]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -1906,20 +1724,16 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5",
-					Usage: "If you already have a socks5, such as 127.0.0.1:1081, and want [src <-> listen socks5 <-> $ brook connect <-> dialWithSocks5 <-> $ brook server/wsserver/wssserver <-> dst]",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Username",
-					Usage: "Optional",
-				},
-				&cli.StringFlag{
-					Name:  "dialWithSocks5Password",
-					Usage: "Optional",
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
 				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook connect --link 'brook://...' --socks5 127.0.0.1:1080")
+					return nil
+				}
 				if c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -1933,19 +1747,18 @@ func main() {
 				if h == "" {
 					return errors.New("socks5 server requires a clear IP for UDP, only port is not enough. You may use loopback IP or lan IP or other, we can not decide for you")
 				}
-				kind, _, _, err := brook.ParseLink(c.String("link"))
+				kind, _, v, err := brook.ParseLink(c.String("link"))
 				if err != nil {
 					return err
 				}
+				if s := v.Get("clientHKDFInfo"); s != "" {
+					brook.ClientHKDFInfo = []byte(s)
+				}
+				if s := v.Get("serverHKDFInfo"); s != "" {
+					brook.ServerHKDFInfo = []byte(s)
+				}
 				if kind == "socks5" {
 					return errors.New("Looks like you want create socks5 from a socks5, you may want $ brook socks5tohttp?")
-				}
-				if c.String("dialWithSocks5") != "" {
-					p, err := socks5dial.NewSocks5Dial(c.String("dialWithSocks5"), c.String("dialWithSocks5Username"), c.String("dialWithSocks5Password"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
-					if err != nil {
-						return err
-					}
-					p.TouchBrook()
 				}
 				s, err := brook.NewBrookLink(c.String("link"))
 				if err != nil {
@@ -1982,7 +1795,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "relay",
-			Usage: "Run as standalone relay, both TCP and UDP, this means access [from address] is equal to access [to address], [src <-> from address <-> to address]",
+			Usage: "Run a standalone relay, which supports TCP and UDP. Accessing [from address] is equal to accessing [to address], [src <-> from address <-> to address]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2010,8 +1823,16 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook relay --from :9999 --to 1.2.3.4:9999")
+					return nil
+				}
 				if c.String("from") == "" || c.String("to") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2032,7 +1853,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "dnsserver",
-			Usage: "Run as standalone dns server, both TCP and UDP",
+			Usage: "Run a standalone dns server",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2047,12 +1868,12 @@ func main() {
 				},
 				&cli.StringFlag{
 					Name:  "dns",
-					Usage: "DNS server which forward to",
+					Usage: "DNS server which forward to. Such as 8.8.8.8:53 or https://dns.google/dns-query?address=8.8.8.8%3A443, the address is required",
 					Value: "8.8.8.8:53",
 				},
 				&cli.StringFlag{
 					Name:  "blockDomainList",
-					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt",
 				},
 				&cli.BoolFlag{
 					Name:  "disableA",
@@ -2072,16 +1893,28 @@ func main() {
 					Value: 60,
 					Usage: "time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook dnsserver --listen 127.0.0.1:53")
+					return nil
+				}
 				if c.String("listen") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
 				if c.String("blockDomainList") != "" && !strings.HasPrefix(c.String("blockDomainList"), "http://") && !strings.HasPrefix(c.String("blockDomainList"), "https://") && !filepath.IsAbs(c.String("blockDomainList")) {
 					return errors.New("--blockDomainList must be with absolute path")
 				}
-				if c.String("blockDomainList") != "" || c.Bool("disableA") || c.Bool("disableAAAA") {
-					p, err := thedns.NewTheDNS(c.String("blockDomainList"), "", "", c.Bool("disableA"), c.Bool("disableAAAA"))
+				var doh string
+				if strings.HasPrefix(c.String("dns"), "https://") {
+					doh = c.String("dns")
+				}
+				if c.String("blockDomainList") != "" || c.Bool("disableA") || c.Bool("disableAAAA") || doh != "" {
+					p, err := thedns.NewTheDNS(c.String("blockDomainList"), "", "", c.Bool("disableA"), c.Bool("disableAAAA"), doh)
 					if err != nil {
 						return err
 					}
@@ -2104,8 +1937,359 @@ func main() {
 			},
 		},
 		&cli.Command{
+			Name:  "dnsclient",
+			Usage: "Send a dns query",
+			BashComplete: func(c *cli.Context) {
+				l := c.Command.VisibleFlags()
+				for _, v := range l {
+					fmt.Println("--" + v.Names()[0])
+				}
+			},
+			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:    "dns",
+					Aliases: []string{"s"},
+					Usage:   "DNS server, such as 8.8.8.8:53",
+					Value:   "8.8.8.8:53",
+				},
+				&cli.StringFlag{
+					Name:    "domain",
+					Aliases: []string{"d"},
+					Usage:   "Domain",
+				},
+				&cli.StringFlag{
+					Name:    "type",
+					Aliases: []string{"t"},
+					Usage:   "Type, such as A",
+					Value:   "A",
+				},
+				&cli.BoolFlag{
+					Name:  "short",
+					Usage: "Short for A/AAAA",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook dnsclient --domain google.com")
+					return nil
+				}
+				if c.String("domain") == "" {
+					return cli.ShowSubcommandHelp(c)
+				}
+				t, ok := dns.StringToType[c.String("type")]
+				if !ok {
+					return errors.New("invalid type")
+				}
+				dc := &brook.DNSClient{Server: c.String("dns")}
+				m := &dns.Msg{}
+				m.SetQuestion(strings.TrimRight(c.String("domain"), ".")+".", t)
+				m, err := dc.Exchange(m)
+				if err != nil {
+					return err
+				}
+				if c.Bool("short") && (c.String("type") == "A" || c.String("type") == "AAAA") {
+					for _, v := range m.Answer {
+						if t, ok := v.(*dns.A); ok {
+							fmt.Println(t.A)
+							return nil
+						}
+						if t, ok := v.(*dns.AAAA); ok {
+							fmt.Println(t.AAAA)
+							return nil
+						}
+					}
+					return nil
+				}
+				fmt.Println(m)
+				return nil
+			},
+		},
+		&cli.Command{
+			Name:  "dohserver",
+			Usage: "Run a standalone doh server",
+			BashComplete: func(c *cli.Context) {
+				l := c.Command.VisibleFlags()
+				for _, v := range l {
+					fmt.Println("--" + v.Names()[0])
+				}
+			},
+			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:  "listen",
+					Usage: "listen address, if you want to create a http server behind nico",
+				},
+				&cli.StringFlag{
+					Name:  "domainaddress",
+					Usage: "Such as: domain.com:443, if you want to create a https server. If you choose to automatically issue certificates, the domain must have been resolved to the server IP and 80 port also will be used",
+				},
+				&cli.StringFlag{
+					Name:  "cert",
+					Usage: "The cert file absolute path for the domain, such as /path/to/cert.pem. If cert or certkey is empty, a certificate will be issued automatically",
+				},
+				&cli.StringFlag{
+					Name:  "certkey",
+					Usage: "The cert key file absolute path for the domain, such as /path/to/certkey.pem. If cert or certkey is empty, a certificate will be issued automatically",
+				},
+				&cli.StringFlag{
+					Name:  "path",
+					Usage: "URL path",
+					Value: "/dns-query",
+				},
+				&cli.StringFlag{
+					Name:  "dns",
+					Usage: "DNS server which forward to. Such as 8.8.8.8:53 or https://dns.google/dns-query?address=8.8.8.8%3A443, the address is required",
+					Value: "8.8.8.8:53",
+				},
+				&cli.StringFlag{
+					Name:  "blockDomainList",
+					Usage: "One domain per line, suffix match mode. https://, http:// or local absolute file path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt",
+				},
+				&cli.BoolFlag{
+					Name:  "disableA",
+					Usage: "Disable A query",
+				},
+				&cli.BoolFlag{
+					Name:  "disableAAAA",
+					Usage: "Disable AAAA query",
+				},
+				&cli.IntFlag{
+					Name:  "tcpTimeout",
+					Value: 0,
+					Usage: "time (s)",
+				},
+				&cli.IntFlag{
+					Name:  "udpTimeout",
+					Value: 60,
+					Usage: "time (s)",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook dohserver --domainaddress domain.com:443")
+					return nil
+				}
+				if c.String("listen") == "" && c.String("domainaddress") == "" {
+					return cli.ShowSubcommandHelp(c)
+				}
+				if c.String("cert") != "" && !filepath.IsAbs(c.String("cert")) {
+					return errors.New("--cert must be with absolute path")
+				}
+				if c.String("certkey") != "" && !filepath.IsAbs(c.String("certkey")) {
+					return errors.New("--certkey must be with absolute path")
+				}
+				if c.String("blockDomainList") != "" && !strings.HasPrefix(c.String("blockDomainList"), "http://") && !strings.HasPrefix(c.String("blockDomainList"), "https://") && !filepath.IsAbs(c.String("blockDomainList")) {
+					return errors.New("--blockDomainList must be with absolute path")
+				}
+				if c.String("blockDomainList") != "" || c.Bool("disableA") || c.Bool("disableAAAA") {
+					p, err := thedns.NewTheDNS(c.String("blockDomainList"), "", "", c.Bool("disableA"), c.Bool("disableAAAA"), "")
+					if err != nil {
+						return err
+					}
+					p.TouchBrook()
+				}
+				addr := ""
+				domain := ""
+				if c.String("domainaddress") != "" {
+					h, p, err := net.SplitHostPort(c.String("domainaddress"))
+					if err != nil {
+						return err
+					}
+					domain = h
+					addr = ":" + p
+				}
+				if c.String("listen") != "" {
+					addr = c.String("listen")
+				}
+				s, err := brook.NewDOHServer(addr, domain, c.String("path"), c.String("dns"), c.Int("tcpTimeout"), c.Int("udpTimeout"))
+				if err != nil {
+					return err
+				}
+				if c.String("cert") != "" {
+					b, err := os.ReadFile(c.String("cert"))
+					if err != nil {
+						return err
+					}
+					s.Cert = b
+				}
+				if c.String("certkey") != "" {
+					b, err := os.ReadFile(c.String("certkey"))
+					if err != nil {
+						return err
+					}
+					s.CertKey = b
+				}
+				g.Add(&runnergroup.Runner{
+					Start: func() error {
+						return s.ListenAndServe()
+					},
+					Stop: func() error {
+						return s.Shutdown()
+					},
+				})
+				return nil
+			},
+		},
+		&cli.Command{
+			Name:  "dohclient",
+			Usage: "Send a dns query",
+			BashComplete: func(c *cli.Context) {
+				l := c.Command.VisibleFlags()
+				for _, v := range l {
+					fmt.Println("--" + v.Names()[0])
+				}
+			},
+			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:    "doh",
+					Aliases: []string{"s"},
+					Usage:   "DOH server, the address is required",
+					Value:   "https://dns.quad9.net/dns-query?address=9.9.9.9%3A443",
+				},
+				&cli.StringFlag{
+					Name:    "domain",
+					Aliases: []string{"d"},
+					Usage:   "Domain",
+				},
+				&cli.StringFlag{
+					Name:    "type",
+					Aliases: []string{"t"},
+					Usage:   "Type, such as A",
+					Value:   "A",
+				},
+				&cli.BoolFlag{
+					Name:  "short",
+					Usage: "Short for A/AAAA",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook dohclient --domain google.com")
+					return nil
+				}
+				if c.String("domain") == "" {
+					return cli.ShowSubcommandHelp(c)
+				}
+				t, ok := dns.StringToType[c.String("type")]
+				if !ok {
+					return errors.New("invalid type")
+				}
+				dc, err := brook.NewDOHClient(c.String("doh"))
+				if err != nil {
+					return err
+				}
+				m := &dns.Msg{}
+				m.SetQuestion(strings.TrimRight(c.String("domain"), ".")+".", t)
+				m, err = dc.Exchange(m)
+				if err != nil {
+					return err
+				}
+				if c.Bool("short") && (c.String("type") == "A" || c.String("type") == "AAAA") {
+					for _, v := range m.Answer {
+						if t, ok := v.(*dns.A); ok {
+							fmt.Println(t.A)
+							return nil
+						}
+						if t, ok := v.(*dns.AAAA); ok {
+							fmt.Println(t.AAAA)
+							return nil
+						}
+					}
+					return nil
+				}
+				fmt.Println(m)
+				return nil
+			},
+		},
+		&cli.Command{
+			Name:  "dhcpserver",
+			Usage: "Run a standalone dhcp server. Other running dhcp servers need to be stopped.",
+			BashComplete: func(c *cli.Context) {
+				l := c.Command.VisibleFlags()
+				for _, v := range l {
+					fmt.Println("--" + v.Names()[0])
+				}
+			},
+			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:  "interface",
+					Usage: "Select interface on multi interface device. Linux only",
+				},
+				&cli.StringFlag{
+					Name:  "serverip",
+					Usage: "DHCP server IP, the IP of the this machine, you shoud set a static IP to this machine before doing this, such as: 192.168.1.10",
+				},
+				&cli.StringFlag{
+					Name:  "start",
+					Usage: "Start IP which you want to assign to clients, such as: 192.168.1.100",
+				},
+				&cli.StringFlag{
+					Name:  "netmask",
+					Usage: "Subnet netmask which you want to assign to clients",
+					Value: "255.255.255.0",
+				},
+				&cli.IntFlag{
+					Name:  "count",
+					Usage: "IP range from the start, which you want to assign to clients",
+					Value: 100,
+				},
+				&cli.StringFlag{
+					Name:  "gateway",
+					Usage: "The router gateway which you want to assign to clients, such as: 192.168.1.1",
+				},
+				&cli.StringSliceFlag{
+					Name:  "dnsserver",
+					Usage: "The dns server which you want to assign to clients, such as: 192.168.1.1 or 8.8.8.8",
+				},
+				&cli.StringFlag{
+					Name:  "cache",
+					Usage: "Cache file, local absolute file path, default is $HOME/.brook.dhcpserver",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook dhcpserver --serverip 192.168.1.10 --start 192.168.1.100 --gateway 192.168.1.1 --dnsserver 192.168.1.1")
+					return nil
+				}
+				if c.String("serverip") == "" || c.String("start") == "" || c.String("gateway") == "" || len(c.StringSlice("dnsserver")) == 0 {
+					return cli.ShowSubcommandHelp(c)
+				}
+				if c.String("cache") != "" && !filepath.IsAbs(c.String("cache")) {
+					return errors.New("--cache must be with absolute path")
+				}
+				s, err := brook.NewDHCPServer(c.String("interface"), c.String("serverip"), c.String("start"), c.String("netmask"), c.Int("count"), c.String("gateway"), c.StringSlice("dnsserver"), c.String("cache"))
+				if err != nil {
+					return err
+				}
+				g.Add(&runnergroup.Runner{
+					Start: func() error {
+						return s.ListenAndServe()
+					},
+					Stop: func() error {
+						return s.Shutdown()
+					},
+				})
+				return nil
+			},
+		},
+		&cli.Command{
 			Name:  "socks5",
-			Usage: "Run as standalone standard socks5 server, both TCP and UDP",
+			Usage: "Run a standalone standard socks5 server, which supports TCP and UDP",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2144,8 +2328,16 @@ func main() {
 					Value: 60,
 					Usage: "Connection deadline time (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook socks5 --listen :9999 --socks5ServerIP 1.2.3.4")
+					return nil
+				}
 				if c.String("listen") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2181,7 +2373,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "socks5tohttp",
-			Usage: "Convert socks5 to http proxy, [src <-> listen address(http proxy) <-> socks5 address <-> dst]",
+			Usage: "Convert a socks5 proxy to a http proxy, [src <-> listen address(http proxy) <-> socks5 address <-> dst]",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2212,8 +2404,16 @@ func main() {
 					Value: 0,
 					Usage: "Connection tcp timeout (s)",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook socks5tohttp --listen 127.0.0.1:8010 --socks5 127.0.0.1:1080")
+					return nil
+				}
 				if c.String("listen") == "" || c.String("socks5") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2234,7 +2434,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "pac",
-			Usage: "Run as PAC server or save PAC to file",
+			Usage: "Run a PAC server or save PAC to a file",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2256,15 +2456,23 @@ func main() {
 				&cli.StringFlag{
 					Name:    "bypassDomainList",
 					Aliases: []string{"b"},
-					Usage:   "One domain per line, suffix match mode. http(s):// or local absolute file path. Like: https://txthinking.github.io/bypass/example_domain.txt",
+					Usage:   "One domain per line, suffix match mode. http(s):// or local absolute file path. Like: https://raw.githubusercontent.com/txthinking/brook/master/programmable/list/example_domain.txt",
 				},
 				&cli.StringFlag{
 					Name:    "file",
 					Aliases: []string{"f"},
 					Usage:   "Save PAC to file, this will ignore listen address",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook pac --listen 127.0.0.1:1980")
+					return nil
+				}
 				if c.String("listen") == "" && c.String("file") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2288,7 +2496,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "testsocks5",
-			Usage: "Test UDP and TCP of socks5 server",
+			Usage: "Test a socks5 server to see if it works properly",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2326,8 +2534,16 @@ func main() {
 					Value: "137.184.237.95",
 					Usage: "The A record of domain",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook testsocks5 --socks5 127.0.0.1:1080")
+					return nil
+				}
 				if c.String("socks5") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2337,7 +2553,7 @@ func main() {
 		},
 		&cli.Command{
 			Name:  "testbrook",
-			Usage: "Test UDP and TCP of brook server/wsserver/wssserver",
+			Usage: "Test UDP and TCP of a brook server/wsserver/wssserver/quicserver connection. (The global-dial-parameter is ignored)",
 			BashComplete: func(c *cli.Context) {
 				l := c.Command.VisibleFlags()
 				for _, v := range l {
@@ -2370,8 +2586,16 @@ func main() {
 					Value: "137.184.237.95",
 					Usage: "The A record of domain",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook testbrook --link 'brook://...'")
+					return nil
+				}
 				if c.String("link") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2385,6 +2609,7 @@ func main() {
 					if err != nil {
 						return
 					}
+					// TODO append global dial options
 					cmd = exec.Command(s, "connect", "--link", c.String("link"), "--socks5", c.String("socks5"))
 					b, _ := cmd.CombinedOutput()
 					err = errors.New(string(b))
@@ -2413,8 +2638,16 @@ func main() {
 					Aliases: []string{"l"},
 					Usage:   "Listen address, like: ':7777'",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook echoserver --listen :9999")
+					return nil
+				}
 				if c.String("listen") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
@@ -2453,12 +2686,56 @@ func main() {
 					Value: 1,
 					Usage: "Times of interactions",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook echoclient --server 1.2.3.4:9999")
+					return nil
+				}
 				if c.String("server") == "" {
 					return cli.ShowSubcommandHelp(c)
 				}
 				return brook.EchoClient(c.String("server"), c.Int("times"))
+			},
+		},
+		&cli.Command{
+			Name:  "ipcountry",
+			Usage: "Get country of IP",
+			BashComplete: func(c *cli.Context) {
+				l := c.Command.VisibleFlags()
+				for _, v := range l {
+					fmt.Println("--" + v.Names()[0])
+				}
+			},
+			Flags: []cli.Flag{
+				&cli.StringFlag{
+					Name:  "ip",
+					Usage: "1.1.1.1",
+				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook ipcountry --ip 1.2.3.4")
+					return nil
+				}
+				ip := net.ParseIP(c.String("ip"))
+				if ip == nil {
+					return errors.New(c.String("ip") + " is not IP")
+				}
+				s := iploc.Country(ip)
+				if s == "" {
+					return errors.New(c.String("ip") + " unknown")
+				}
+				fmt.Println(s)
+				return nil
 			},
 		},
 		&cli.Command{
@@ -2471,8 +2748,16 @@ func main() {
 					Usage:   "Write to file",
 					Value:   "brook_autocomplete",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook completion --file brook_autocomplete")
+					return nil
+				}
 				l := c.App.VisibleCommands()
 				if strings.Contains(os.Getenv("SHELL"), "zsh") {
 					s := `
@@ -2560,8 +2845,16 @@ complete -o bashdefault -o default -o nospace -F _cli_bash_autocomplete brook
 					Aliases: []string{"f"},
 					Usage:   "Write to file, default print to stdout",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook mdpage")
+					return nil
+				}
 				s, err := c.App.ToMarkdown()
 				if err != nil {
 					return err
@@ -2591,8 +2884,16 @@ complete -o bashdefault -o default -o nospace -F _cli_bash_autocomplete brook
 					Aliases: []string{"f"},
 					Usage:   "Write to file, default print to stdout. You should put to /path/to/man/man1/brook.1 on linux or /usr/local/share/man/man1/brook.1 on macos",
 				},
+				&cli.BoolFlag{
+					Name:  "example",
+					Usage: "Show a minimal example of usage",
+				},
 			},
 			Action: func(c *cli.Context) error {
+				if c.Bool("example") {
+					fmt.Println("brook manpage")
+					return nil
+				}
 				s, err := c.App.ToManWithSection(1)
 				if err != nil {
 					return err
@@ -2611,11 +2912,37 @@ complete -o bashdefault -o default -o nospace -F _cli_bash_autocomplete brook
 	if os.Getenv("SOCKS5_DEBUG") != "" {
 		socks5.Debug = true
 	}
-	if err := app.Run(os.Args); err != nil {
+	l := os.Args
+	if len(os.Args) == 2 {
+		if _, err := os.Stat(os.Args[1]); err == nil {
+			if !filepath.IsAbs(os.Args[1]) {
+				log.Println("It looks like you passed config file? It must be absolute path")
+				os.Exit(1)
+				return
+			}
+			l, err = brook.CAC(os.Args[1])
+			if err != nil {
+				log.Println(err)
+				os.Exit(1)
+				return
+			}
+			bin, err := os.Executable()
+			if err != nil {
+				log.Println(err)
+				os.Exit(1)
+				return
+			}
+			l = append([]string{bin}, l...)
+		}
+	}
+	if err := app.Run(l); err != nil {
 		log.Println(err)
+		df()
+		os.Exit(1)
 		return
 	}
 	if len(g.Runners) == 0 {
+		df()
 		return
 	}
 	go func() {
@@ -2624,5 +2951,11 @@ complete -o bashdefault -o default -o nospace -F _cli_bash_autocomplete brook
 		<-sigs
 		g.Done()
 	}()
-	log.Println(g.Wait())
+	if err := g.Wait(); err != nil {
+		log.Println(err)
+		df()
+		os.Exit(1)
+		return
+	}
+	df()
 }
