@@ -18,22 +18,32 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
+	"io"
 	"net"
 	"net/url"
+	"os/exec"
+	"runtime"
+	"strconv"
+	"strings"
 
+	utls "github.com/refraction-networking/utls"
 	"github.com/txthinking/brook/limits"
-	crypto1 "github.com/txthinking/crypto"
 	"github.com/txthinking/socks5"
 )
 
 type BrookLink struct {
-	Kind     string
-	Address  string
-	Host     string
-	Path     string
-	Password []byte
-	V        url.Values
-	Tc       *tls.Config
+	Kind              string
+	Address           string
+	Host              string
+	Path              string
+	Password          []byte
+	V                 url.Values
+	Tc                *tls.Config
+	TLSFingerprint    utls.ClientHelloID
+	FragmentMinLength int64
+	FragmentMaxLength int64
+	FragmentMinDelay  int64
+	FragmentMaxDelay  int64
 
 	S5         *socks5.Server
 	Pcf        *PacketConnFactory
@@ -52,6 +62,11 @@ func NewBrookLink(link string) (*BrookLink, error) {
 		address = server
 	}
 	var tc *tls.Config
+	var tlsfingerprint utls.ClientHelloID
+	var fragmentMinLength int64
+	var fragmentMaxLength int64
+	var fragmentMinDelay int64
+	var fragmentMaxDelay int64
 	if kind == "socks5" || kind == "wsserver" || kind == "wssserver" || kind == "quicserver" {
 		u, err := url.Parse(server)
 		if err != nil {
@@ -85,73 +100,94 @@ func NewBrookLink(link string) (*BrookLink, error) {
 				}
 				tc.RootCAs = roots
 			}
+			if kind == "wssserver" {
+				tc.NextProtos = []string{"http/1.1"}
+			}
 			if kind == "quicserver" {
 				tc.NextProtos = []string{"h3"}
 			}
 		}
 		if kind == "wsserver" || kind == "wssserver" || kind == "quicserver" {
 			if v.Get("withoutBrookProtocol") == "true" {
-				p, err = crypto1.SHA256Bytes([]byte(v.Get("password")))
+				p, err = SHA256Bytes([]byte(v.Get("password")))
 				if err != nil {
 					return nil, err
 				}
 			}
 		}
+		if kind == "wssserver" {
+			if v.Get("tlsfingerprint") == "chrome" {
+				tlsfingerprint = utls.HelloChrome_Auto
+			}
+			if v.Get("fragment") != "" {
+				l := strings.Split(v.Get("fragment"), ":")
+				if len(l) == 4 {
+					fragmentMinLength, _ = strconv.ParseInt(l[0], 10, 64)
+					fragmentMaxLength, _ = strconv.ParseInt(l[1], 10, 64)
+					fragmentMinDelay, _ = strconv.ParseInt(l[2], 10, 64)
+					fragmentMaxDelay, _ = strconv.ParseInt(l[3], 10, 64)
+				}
+			}
+		}
 	}
 	return &BrookLink{
-		Kind:     kind,
-		Address:  address,
-		Host:     host,
-		Path:     path,
-		Password: p,
-		V:        v,
-		Tc:       tc,
+		Kind:              kind,
+		Address:           address,
+		Host:              host,
+		Path:              path,
+		Password:          p,
+		V:                 v,
+		Tc:                tc,
+		TLSFingerprint:    tlsfingerprint,
+		FragmentMinLength: fragmentMinLength,
+		FragmentMaxLength: fragmentMaxLength,
+		FragmentMinDelay:  fragmentMinDelay,
+		FragmentMaxDelay:  fragmentMaxDelay,
 	}, nil
 }
 
-// The caller is responsible for closing the connection in Exchanger.
-func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeout, udptimeout int) (Exchanger, error) {
+func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeout, udptimeout int) (Exchanger, net.Conn, error) {
 	if blk.Kind == "server" {
 		if network == "tcp" {
 			rc, err := DialTCP("tcp", "", blk.Address)
 			if err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 			sc, err := NewStreamClient("tcp", blk.Password, src, rc, tcptimeout, dstb)
 			if err != nil {
 				rc.Close()
-				return nil, err
+				return nil, nil, err
 			}
-			return sc, nil
+			return sc, rc, nil
 		}
 		if blk.V.Get("udpovertcp") == "true" {
 			rc, err := NATDial("tcp", src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address)
 			if err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 			sc, err := NewStreamClient("udp", blk.Password, src, rc, udptimeout, dstb)
 			if err != nil {
 				rc.Close()
-				return nil, err
+				return nil, nil, err
 			}
-			return sc, nil
+			return sc, rc, nil
 		}
 		rc, err := NATDial("udp", src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address)
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 		sc, err := NewPacketClient(blk.Password, src, rc, udptimeout, dstb)
 		if err != nil {
 			rc.Close()
-			return nil, err
+			return nil, nil, err
 		}
-		return sc, nil
+		return sc, rc, nil
 	}
 	if blk.Kind == "wsserver" || blk.Kind == "wssserver" {
 		if network == "tcp" {
-			rc, err := WebSocketDial("", "", blk.Address, blk.Host, blk.Path, blk.Tc, tcptimeout)
+			rc, err := WebSocketDial("", "", blk.Address, blk.Host, blk.Path, blk.Tc, tcptimeout, blk.TLSFingerprint, blk.FragmentMinLength, blk.FragmentMaxLength, blk.FragmentMinDelay, blk.FragmentMaxDelay)
 			if err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 			var sc Exchanger
 			if blk.V.Get("withoutBrookProtocol") != "true" {
@@ -162,13 +198,13 @@ func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeo
 			}
 			if err != nil {
 				rc.Close()
-				return nil, err
+				return nil, nil, err
 			}
-			return sc, nil
+			return sc, rc, nil
 		}
-		rc, err := WebSocketDial(src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address, blk.Host, blk.Path, blk.Tc, tcptimeout)
+		rc, err := WebSocketDial(src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address, blk.Host, blk.Path, blk.Tc, tcptimeout, blk.TLSFingerprint, blk.FragmentMinLength, blk.FragmentMaxLength, blk.FragmentMinDelay, blk.FragmentMaxDelay)
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 		var sc Exchanger
 		if blk.V.Get("withoutBrookProtocol") != "true" {
@@ -179,15 +215,15 @@ func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeo
 		}
 		if err != nil {
 			rc.Close()
-			return nil, err
+			return nil, nil, err
 		}
-		return sc, nil
+		return sc, rc, nil
 	}
 	if blk.Kind == "quicserver" {
 		if network == "tcp" {
-			rc, err := QUICDialTCP("", "", blk.Address, blk.Host, blk.Tc, tcptimeout)
+			rc, err := QUICDialTCP("", "", blk.Address, blk.Tc, tcptimeout)
 			if err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 			var sc Exchanger
 			if blk.V.Get("withoutBrookProtocol") != "true" {
@@ -198,13 +234,31 @@ func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeo
 			}
 			if err != nil {
 				rc.Close()
-				return nil, err
+				return nil, nil, err
 			}
-			return sc, nil
+			return sc, rc, nil
 		}
-		rc, err := QUICDialUDP(src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address, blk.Host, blk.Tc, udptimeout)
+		if blk.V.Get("udpoverstream") == "true" {
+			rc, err := QUICDialTCP(src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address, blk.Tc, tcptimeout)
+			if err != nil {
+				return nil, nil, err
+			}
+			var sc Exchanger
+			if blk.V.Get("withoutBrookProtocol") != "true" {
+				sc, err = NewStreamClient("udp", blk.Password, src, rc, tcptimeout, dstb)
+			}
+			if blk.V.Get("withoutBrookProtocol") == "true" {
+				sc, err = NewSimpleStreamClient("udp", blk.Password, src, rc, tcptimeout, dstb)
+			}
+			if err != nil {
+				rc.Close()
+				return nil, nil, err
+			}
+			return sc, rc, nil
+		}
+		rc, err := QUICDialUDP(src, socks5.ToAddress(dstb[0], dstb[1:len(dstb)-2], dstb[len(dstb)-2:]), blk.Address, blk.Tc, udptimeout)
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 		var sc Exchanger
 		if blk.V.Get("withoutBrookProtocol") != "true" {
@@ -215,21 +269,37 @@ func (blk *BrookLink) CreateExchanger(network, src string, dstb []byte, tcptimeo
 		}
 		if err != nil {
 			rc.Close()
-			return nil, err
+			return nil, nil, err
 		}
-		return sc, nil
+		return sc, rc, nil
 	}
-	return nil, errors.New("cannot create exchanger from " + blk.Kind)
+	return nil, nil, errors.New("cannot create exchanger from " + blk.Kind)
 }
 
 func (x *BrookLink) PrepareSocks5Server(addr, ip string, tcptimeout, udptimeout int) error {
+	if err := limits.Raise(); err != nil {
+		Log(Error{"when": "try to raise system limits", "warning": err.Error()})
+	}
+	if x.Kind == "quicserver" {
+		if runtime.GOOS == "linux" {
+			c := exec.Command("sysctl", "-w", "net.core.rmem_max=2500000")
+			b, err := c.CombinedOutput()
+			if err != nil {
+				Log(Error{"when": "try to raise UDP Receive Buffer Size", "warning": string(b)})
+			}
+		}
+		if runtime.GOOS == "darwin" {
+			c := exec.Command("sysctl", "-w", "kern.ipc.maxsockbuf=3014656")
+			b, err := c.CombinedOutput()
+			if err != nil {
+				Log(Error{"when": "try to raise UDP Receive Buffer Size", "warning": string(b)})
+			}
+		}
+	}
 	var err error
 	x.S5, err = socks5.NewClassicServer(addr, ip, "", "", tcptimeout, udptimeout)
 	if err != nil {
 		return err
-	}
-	if err := limits.Raise(); err != nil {
-		Log(&Error{"when": "try to raise system limits", "warning": err.Error()})
 	}
 	x.Pcf = NewPacketConnFactory()
 	x.TCPTimeout = tcptimeout
@@ -244,21 +314,13 @@ func (x *BrookLink) ListenAndServe() error {
 func (x *BrookLink) TCPHandle(s *socks5.Server, c *net.TCPConn, r *socks5.Request) error {
 	if r.Cmd == socks5.CmdConnect {
 		dstb := append(append([]byte{r.Atyp}, r.DstAddr...), r.DstPort...)
-		sc, err := x.CreateExchanger("tcp", c.RemoteAddr().String(), dstb, x.TCPTimeout, x.UDPTimeout)
+		sc, rc, err := x.CreateExchanger("tcp", c.RemoteAddr().String(), dstb, x.TCPTimeout, x.UDPTimeout)
 		if err != nil {
 			return ErrorReply(r, c, err)
 		}
+		defer rc.Close()
 		defer sc.Clean()
-		laddr := ""
-		if v, ok := sc.(*StreamClient); ok {
-			defer v.Server.Close()
-			laddr = v.Server.LocalAddr().String()
-		}
-		if v, ok := sc.(*SimpleStreamClient); ok {
-			defer v.Server.Close()
-			laddr = v.Server.LocalAddr().String()
-		}
-		a, address, port, err := socks5.ParseAddress(laddr)
+		a, address, port, err := socks5.ParseAddress(rc.LocalAddr().String())
 		if err != nil {
 			return ErrorReply(r, c, err)
 		}
@@ -276,6 +338,7 @@ func (x *BrookLink) TCPHandle(s *socks5.Server, c *net.TCPConn, r *socks5.Reques
 		if err != nil {
 			return err
 		}
+		io.Copy(io.Discard, c)
 		return nil
 	}
 	return socks5.ErrUnsupportCmd
@@ -294,23 +357,12 @@ func (x *BrookLink) UDPHandle(s *socks5.Server, addr *net.UDPAddr, d *socks5.Dat
 		return nil
 	}
 	defer conn.Close()
-	sc, err := x.CreateExchanger("udp", addr.String(), dstb, x.TCPTimeout, x.UDPTimeout)
+	sc, rc, err := x.CreateExchanger("udp", addr.String(), dstb, x.TCPTimeout, x.UDPTimeout)
 	if err != nil {
 		return err
 	}
+	defer rc.Close()
 	defer sc.Clean()
-	if v, ok := sc.(*PacketClient); ok {
-		defer v.Server.Close()
-	}
-	if v, ok := sc.(*StreamClient); ok {
-		defer v.Server.Close()
-	}
-	if v, ok := sc.(*SimplePacketClient); ok {
-		defer v.Server.Close()
-	}
-	if v, ok := sc.(*SimpleStreamClient); ok {
-		defer v.Server.Close()
-	}
 	if err := sc.Exchange(conn); err != nil {
 		return nil
 	}
