@@ -65,7 +65,7 @@ func (c *PacketClient) Exchange(local net.Conn) error {
 				return
 			}
 			sk := x.BP32.Get().([]byte)
-			if _, err := io.ReadFull(hkdf.New(sha256.New, c.Password, c.RB[:12], []byte{0x62, 0x72, 0x6f, 0x6f, 0x6b}), sk); err != nil {
+			if _, err := io.ReadFull(hkdf.New(sha256.New, c.Password, c.RB[:12], ServerHKDFInfo), sk); err != nil {
 				x.BP32.Put(sk)
 				Log(err)
 				return
@@ -113,7 +113,7 @@ func (c *PacketClient) Exchange(local net.Conn) error {
 		}
 		binary.BigEndian.PutUint32(c.WB[12:12+4], uint32(time.Now().Unix()))
 		ck := x.BP32.Get().([]byte)
-		if _, err := io.ReadFull(hkdf.New(sha256.New, c.Password, c.WB[:12], []byte{0x62, 0x72, 0x6f, 0x6f, 0x6b}), ck); err != nil {
+		if _, err := io.ReadFull(hkdf.New(sha256.New, c.Password, c.WB[:12], ClientHKDFInfo), ck); err != nil {
 			x.BP32.Put(ck)
 			return err
 		}
