@@ -64,7 +64,7 @@ func (s *PacketServer) Exchange(remote net.Conn) error {
 				return
 			}
 			sk := x.BP32.Get().([]byte)
-			if _, err := io.ReadFull(hkdf.New(sha256.New, s.Password, s.WB[:12], []byte{0x62, 0x72, 0x6f, 0x6f, 0x6b}), sk); err != nil {
+			if _, err := io.ReadFull(hkdf.New(sha256.New, s.Password, s.WB[:12], ServerHKDFInfo), sk); err != nil {
 				x.BP32.Put(sk)
 				Log(err)
 				return
