@@ -20,6 +20,7 @@ import (
 	"crypto/sha256"
 	"encoding/binary"
 	"errors"
+	"fmt"
 	"io"
 	"net"
 	"sync"
@@ -47,7 +48,7 @@ func (f *PacketServerConnFactory) Handle(addr *net.UDPAddr, b, p []byte, w func(
 		return nil, nil, errors.New("data too small")
 	}
 	ck := x.BP32.Get().([]byte)
-	if _, err := io.ReadFull(hkdf.New(sha256.New, p, b[:12], []byte{0x62, 0x72, 0x6f, 0x6f, 0x6b}), ck); err != nil {
+	if _, err := io.ReadFull(hkdf.New(sha256.New, p, b[:12], ClientHKDFInfo), ck); err != nil {
 		x.BP32.Put(ck)
 		return nil, nil, err
 	}
@@ -72,6 +73,9 @@ func (f *PacketServerConnFactory) Handle(addr *net.UDPAddr, b, p []byte, w func(
 	a, h, p, err := socks5.ParseBytesAddress(b[12+4:])
 	if err != nil {
 		return nil, nil, err
+	}
+	if 12+4+1+len(h)+2 >= len(b)-16 {
+		return nil, nil, errors.New(fmt.Sprintf("invalid packet. length: %d address: %#x %#x %#x", len(b), a, h, p))
 	}
 	dst := socks5.ToAddress(a, h, p)
 	f.Lock.Lock()
