@@ -23,7 +23,7 @@ import (
 	"github.com/quic-go/quic-go"
 )
 
-func QUICDialUDP(src, dst, addr, host string, tc *tls.Config, idleTime int) (net.Conn, error) {
+func QUICDialUDP(src, dst, addr string, tc *tls.Config, idleTime int) (net.Conn, error) {
 	var rc *net.UDPConn
 	var err error
 	if src == "" || dst == "" {
@@ -40,7 +40,7 @@ func QUICDialUDP(src, dst, addr, host string, tc *tls.Config, idleTime int) (net
 		rc.Close()
 		return nil, err
 	}
-	rc1, err := quic.Dial(rc, raddr.(*net.UDPAddr), host, tc, &quic.Config{MaxIdleTimeout: time.Duration(idleTime) * time.Second, EnableDatagrams: true})
+	rc1, err := quic.Dial(context.Background(), rc, raddr, tc, &quic.Config{MaxIdleTimeout: time.Duration(idleTime) * time.Second, EnableDatagrams: true})
 	if err != nil {
 		rc.Close()
 		return nil, err
@@ -53,7 +53,7 @@ func QUICDialUDP(src, dst, addr, host string, tc *tls.Config, idleTime int) (net
 	}, nil
 }
 
-func QUICDialTCP(src, dst, addr, host string, tc *tls.Config, idleTime int) (net.Conn, error) {
+func QUICDialTCP(src, dst, addr string, tc *tls.Config, idleTime int) (net.Conn, error) {
 	var rc *net.UDPConn
 	var err error
 	if src == "" || dst == "" {
@@ -70,7 +70,7 @@ func QUICDialTCP(src, dst, addr, host string, tc *tls.Config, idleTime int) (net
 		rc.Close()
 		return nil, err
 	}
-	rc1, err := quic.Dial(rc, raddr.(*net.UDPAddr), host, tc, &quic.Config{MaxIdleTimeout: time.Duration(idleTime) * time.Second})
+	rc1, err := quic.Dial(context.Background(), rc, raddr, tc, &quic.Config{MaxIdleTimeout: time.Duration(idleTime) * time.Second})
 	if err != nil {
 		rc.Close()
 		return nil, err
@@ -110,7 +110,7 @@ func (c *QUICConn) Read(b []byte) (int, error) {
 	if c.Stream != nil {
 		return c.Stream.Read(b)
 	}
-	b1, err := c.Conn.ReceiveMessage()
+	b1, err := c.Conn.ReceiveDatagram(context.Background())
 	if err != nil {
 		return 0, err
 	}
@@ -122,7 +122,7 @@ func (c *QUICConn) Write(b []byte) (int, error) {
 	if c.Stream != nil {
 		return c.Stream.Write(b)
 	}
-	if err := c.Conn.SendMessage(b); err != nil {
+	if err := c.Conn.SendDatagram(b); err != nil {
 		return 0, err
 	}
 	return len(b), nil
